extender move server and sni files out of core
We don't want to package random server names in our binary.
Use compressed sni files to load in server names.
