+++ extender/.gitignore
@@ -1,0 +1,4 @@
+
+# ignore the raw files. These are local for editing
+sni-mail
+sni-service
