+++ net_extender_profiles.go
@@ -80,4410 +80,8 @@ var MailPorts = map[int][]ExtenderConnectMode{
 	465: []ExtenderConnectMode{ExtenderConnectModeTcpTls},
 }
 
-var serviceHosts = []string{
-	"github.com",
-	"amazonaws.com",
-	"googleapis.com",
-	"npmjs.com",
-	"nuget.org",
-	"dockerhub.com",
-	"maven.org",
-	"rubygems.org",
-	"packagist.org",
-	"pypi.org",
-	"debian.org",
-	"centos.org",
-	"fedoraproject.org",
-	"registry.npmjs.org",
-	"hub.docker.com",
-	"gcr.io",
-	"quay.io",
-	"registry.gitlab.com",
-	"conan.io",
-	"bintray.com",
-	"jfrog.io",
-	"pkg.go.dev",
-	"pub.dev",
-	"crates.io",
-	"cocoapods.org",
-	"hackage.haskell.org",
-	"hex.pm",
-	"jitpack.io",
-	"packagecloud.io",
-	"openjdk.java.net",
-	"dl.bintray.com",
-	"repo.maven.apache.org",
-	"download.eclipse.org",
-	"registry.yarnpkg.com",
-	"plugins.gradle.org",
-	"search.maven.org",
-	"clojars.org",
-	"repo1.maven.org",
-	"repo.continuum.io",
-	"conda.anaconda.org",
-	"bioconda.github.io",
-	"conda-forge.org",
-	"cran.r-project.org",
-	"bioconductor.org",
-	"deb.nodesource.com",
-	"developer.apple.com",
-	"brew.sh",
-	"dl.google.com",
-	"storage.googleapis.com",
-	"ppa.launchpad.net",
-	"download.opensuse.org",
-	"rpm.nodesource.com",
-	"releases.hashicorp.com",
-	"repo.saltstack.com",
-	"packages.chef.io",
-	"downloads.mesosphere.com",
-	"downloads.sourceforge.net",
-	"get.jenkins.io",
-	"updates.jenkins-ci.org",
-	"download.jenkins.io",
-	"services.gradle.org",
-	"plugins.jetbrains.com",
-	"repo.continuum.io",
-	"repo.hortonworks.com",
-	"mirror.centos.org",
-	"apt.postgresql.org",
-	"download.elastic.co",
-	"artifacts.elastic.co",
-	"repo.zabbix.com",
-	"yum.postgresql.org",
-	"download.docker.com",
-	"packages.gitlab.com",
-	"apt.kubernetes.io",
-	"dl.yarnpkg.com",
-	"download.mono-project.com",
-	"files.pythonhosted.org",
-	"packagecloud.io",
-	"downloads.raspberrypi.org",
-	"repo.mysql.com",
-	"archive.canonical.com",
-	"repository.jboss.org",
-	"dl-ssl.google.com",
-	"repo.mongodb.org",
-	"dl.influxdata.com",
-	"repos.influxdata.com",
-	"downloads.mongodb.com",
-	"packagecloud.io",
-	"packagecloud.io",
-	"packagecloud.io",
-	"downloads.erlang-solutions.com",
-	"repositories.intel.com",
-	"download.qt.io",
-	"dist.scala-sbt.org",
-	"download.jboss.org",
-	"downloads.lightbend.com",
-	"get.docker.com",
-	"downloads.chef.io",
-	"downloads.asterisk.org",
-	"binaries.sonarsource.com",
-	"packages.microsoft.com",
-	"apt.llvm.org",
-	"downloads.mariadb.com",
-	"download.java.net",
-	"binary.plantuml.com",
-	"toolbelt.heroku.com",
-	"download.jetbrains.com",
-	"public.dhe.ibm.com",
-	"repo.spring.io",
-	"repo.typesafe.com",
-	"launchpad.net",
-	"downloads.atlassian.com",
-	"packagecloud.io",
-	"downloads.buildpacks.io",
-	"pkg.jenkins.io",
-	"alpine.global.ssl.fastly.net",
-	"download.gnome.org",
-	"rpmfind.net",
-	"repo.percona.com",
-	"repositories.confluent.io",
-	"repos.azulsystems.com",
-	"apt.kitware.com",
-	"dl.fedoraproject.org",
-	"linuxmint.com",
-	"download.nus.edu.sg",
-	"packages.vmware.com",
-	"downloads.linux.hpe.com",
-	"download.osgeo.org",
-	"packages.cloud.google.com",
-	"downloads.sourceforge.net",
-	"cdn.mysql.com",
-	"cdn.redhat.com",
-	"mirror.centos.org",
-	"download-ib01.fedoraproject.org",
-	"pkg.cfssl.org",
-	"rpmrepo.spacewalk.server",
-	"download.proxmox.com",
-	"repos.mesosphere.com",
-	"bin.equinox.io",
-	"repos.fedorapeople.org",
-	"downloads-distro.mongodb.org",
-	"packages.wazuh.com",
-	"repo.varnish-cache.org",
-	"download.geofabrik.de",
-	"apt.releases.hashicorp.com",
-	"download.geonames.org",
-	"packages.erlang-solutions.com",
-	"download.oracle.com",
-	"rpmrepo.kaltura.org",
-	"repo.huaweicloud.com",
-	"downloads.intercomcdn.com",
-	"downloads.metasploit.com",
-	"download.freebsd.org",
-	"downloads.slack-edge.com",
-	"ftp5.gwdg.de",
-	"packages.treasuredata.com",
-	"updates.signal.org",
-	"repo.continuum.io",
-	"downloads.xiph.org",
-	"packages.endpoint.com",
-	"downloads.sonatype.com",
-	"download.blender.org",
-	"binaries.erlang-solutions.com",
-	"download.virtualbox.org",
-	"downloads.services.jetbrains.com",
-	"downloads.splunk.com",
-	"get.pulumi.com",
-	"downloads.npmjs.org",
-	"pkgs.dyn.su.se",
-	"downloads.vespa.ai",
-	"downloads.graylog2.org",
-	"downloads.tryton.org",
-	"downloads.npcglib.org",
-	"download.syncthing.net",
-	"repo1.maven.org",
-	"download.java.net",
-	"updates.signal.org",
-	"downloads.puppetlabs.com",
-	"pkg.jenkins-ci.org",
-	"release.axocdn.com",
-	"repos.mia.pivotal.io",
-	"yum.puppetlabs.com",
-	"download.pytorch.org",
-	"go.mongodb.org",
-	"downloads.eggheads.net",
-	"dl.k8s.io",
-	"download.splunk.com",
-	"binaries-cdn.jetbrains.com",
-	"cdn.cocoapods.org",
-	"downloads.typesafe.com",
-	"download.ocamlcore.org",
-	"download.eclipse.org",
-	"download.jboss.org",
-	"download.visualstudio.microsoft.com",
-	"pkg.cloudfoundry.org",
-	"downloads.openwrt.org",
-	"download.gluster.org",
-	"packages.fluentbit.io",
-	"repo.huaweicloud.com",
-	"download.jetbrains.com",
-	"download.jetbrains.com",
+// FIXME load from a bit masked sni-service.gz. We don't want any url fragments to be in the binary
+var serviceHosts = []string{}
 
-	"akamaihd.net",
-	"cloudfront.net",
-	"cloudflare.com",
-	"googleusercontent.com",
-	"fbcdn.net",
-	"edgekey.net",
-	"llnwd.net",
-	"edgesuite.net",
-	"amazonaws.com",
-	"cdninstagram.com",
-	"microsoft.com",
-	"hwcdn.net",
-	"yottaa.net",
-	"azureedge.net",
-	"wp.com",
-	"cloudflare.net",
-	"panthercdn.com",
-	"jsdelivr.net",
-	"incapdns.net",
-	"bitgravity.com",
-	"cachefly.net",
-	"cdn77.net",
-	"netdna-cdn.com",
-	"akamai.net",
-	"gstatic.com",
-	"fastly.net",
-	"vo.msecnd.net",
-	"footprint.net",
-	"ay1.b.yahoo.com",
-	"google.com",
-	"mcafee.com",
-	"cdnads.com",
-	"kxcdn.com",
-	"nyiftw.net",
-	"instacontent.net",
-	"twimg.com",
-	"ytimg.com",
-	"msecnd.net",
-	"bcst.cdn.hbo.com",
-	"vimeocdn.com",
-	"keycdn.com",
-	"ssl-images-amazon.com",
-	"netflix.com",
-	"inscname.net",
-	"clients.turbobytes.net",
-	"yahooapis.com",
-	"dropboxstatic.com",
-	"voxcdn.com",
-	"pressable.com",
-	"cedexis.net",
-	"gslb.taobao.com",
-	"squixa.net",
-	"cdn.dimcloud.net",
-	"mncdn.com",
-	"streamroot.io",
-	"alicdn.com",
-	"ngenix.net",
-	"360wzb.com",
-	"sirv.com",
-	"awsdns-53.net",
-	"cdnetworks.net",
-	"zenedge.net",
-	"cdngc.net",
-	"kxcdn.com",
-	"rapidvideo.com",
-	"imgix.net",
-	"nocookie.net",
-	"cdn-apple.com",
-	"apester.com",
-	"distilnetworks.com",
-	"cloudfront.net",
-	"maxcdn.com",
-	"speedcdns.com",
-	"webzdarma.cz",
-	"qingcdn.com",
-	"turbobytes-cdn.com",
-	"videodelivery.net",
-	"insnw.net",
-	"lswcdn.net",
-	"swiftserve.com",
-	"nflxvideo.net",
-	"azurewebsites.net",
-	"d1bpn0ddgcnk2t.cloudfront.net",
-	"theglobeandmail.com",
-	"azureedge.com",
-	"gcdn.co",
-	"cloudinary.com",
-	"doubleverify.com",
-	"nflximg.net",
-	"pinterest.com",
-	"seagate.com",
-	"phncdn.com",
-	"upsellit.com",
-	"criteo.net",
-	"dnvr-vod-cdn.uplynk.com",
-	"rncdn1.com",
-	"jsdelivr.net",
-	"g.doubleclick.net",
-	"imrworldwide.com",
-	"images-amazon.com",
-	"xboxlive.com",
-	"cedexis-radar.net",
-	"scene7.com",
-	"ttvnw.net",
-	"nflxext.com",
-	"aaplimg.com",
-	"digitaloceanspaces.com",
-	"edgecastcdn.net",
-	"delvenetworks.com",
-	"wscdns.com",
-	"videopress.com",
-	"transactcdn.com",
-	"sstatic.net",
-	"parastorage.com",
-	"static-netflix.net",
-	"verizon.net",
-	"oclc.org",
-	"stackassets.com",
-	"att-idns.net",
-	"openx.net",
-	"azure.net",
-	"redditstatic.com",
-	"azure-api.net",
-	"static-nytimes.com",
-	"snapwidget.com",
-	"zenfs.com",
-	"yimg.jp",
-	"steamcontent.com",
-	"rackcdn.com",
-	"verizonwireless.com",
-	"atlassian.com",
-	"yahooapis.com",
-	"shoppilot.com",
-	"staticworld.net",
-	"steamstatic.com",
-	"tribalfusion.com",
-	"apple-dns.net",
-	"imageshack.us",
-	"tenor.co",
-	"dynect.net",
-	"tubemogul.com",
-	"crwdcntrl.net",
-	"2o7.net",
-	"omtrdc.net",
-	"brightcove.com",
-	"optimizely.com",
-	"quantserve.com",
-	"cloudapp.net",
-	"zencdn.net",
-	"mirror-image.net",
-	"bluekai.com",
-	"trafficshield.com",
-	"voxel.net",
-	"gravatar.com",
-	"wsimg.com",
-	"serving-sys.com",
-	"akamai.com",
-	"rubiconproject.com",
-	"static-economist.com",
-	"ebayimg.com",
-	"yumenetworks.com",
-	"gumgum.com",
-	"gigya.com",
-	"ebayrtm.com",
-	"bkrtx.com",
-	"tumblr.com",
-	"tapad.com",
-	"turn.com",
-	"lijit.com",
-	"nr-data.net",
-	"adap.tv",
-	"mookie1.com",
-	"tiqcdn.com",
-	"cedexis.com",
-	"livefyre.com",
-	"turner.com",
-	"ggpht.com",
-	"ads-twitter.com",
-	"nflxso.net",
-	"tremorhub.com",
-	"apnewsregistry.com",
-	"nexac.com",
-	"aolcdn.com",
-	"advertising.com",
-	"adblade.com",
-	"rfihub.com",
-	"undertone.com",
-	"outbrain.com",
-	"betrad.com",
-	"newrelic.com",
-	"pardot.com",
-	"vidible.tv",
-	"po.st",
-	"digitalrivercontent.net",
-	"scorecardresearch.com",
-	"brightcovecdn.com",
-	"atdmt.com",
-	"demdex.net",
-	"photobucket.com",
-	"fyre.co",
-
-	"jpmorgan.com",
-	"bankofamerica.com",
-	"wellsfargo.com",
-	"citigroup.com",
-	"hsbc.com",
-	"goldmansachs.com",
-	"morganstanley.com",
-	"barclays.co.uk",
-	"ubs.com",
-	"creditsuisse.com",
-	"deutschebank.com",
-	"bnpparibas.com",
-	"santander.com",
-	"societegenerale.com",
-	"ing.com",
-	"unicredit.eu",
-	"bbva.com",
-	"standardchartered.com",
-	"usbank.com",
-	"pnc.com",
-	"truist.com",
-	"capitalone.com",
-	"ally.com",
-	"americanexpress.com",
-	"discover.com",
-	"synchrony.com",
-	"schwab.com",
-	"fidelity.com",
-	"vanguard.com",
-	"blackrock.com",
-	"statestreet.com",
-	"northerntrust.com",
-	"bnymellon.com",
-	"invesco.com",
-	"franklintempleton.com",
-	"troweprice.com",
-	"amundi.com",
-	"schroders.com",
-	"allianzgi.com",
-	"pimco.com",
-	"wellingtonfunds.com",
-	"mfs.com",
-	"putnam.com",
-	"columbiathreadneedleus.com",
-	"janushenderson.com",
-	"pgim.com",
-	"lordabbett.com",
-	"nb.com",
-	"leggmason.com",
-	"dodge-cox.com",
-	"americancentury.com",
-	"thriventfunds.com",
-	"hartfordfunds.com",
-	"jhannuities.com",
-	"massmutual.com",
-	"newyorklife.com",
-	"prudential.com",
-	"metlife.com",
-	"aiglife.com",
-	"aflac.com",
-	"nationwide.com",
-	"statefarm.com",
-	"allstate.com",
-	"libertymutual.com",
-	"travelers.com",
-	"progressive.com",
-	"chubb.com",
-	"thehartford.com",
-	"lincoln.com",
-	"guardian.com",
-	"pacificlife.com",
-	"protective.com",
-	"unioncentral.com",
-	"cnoinc.com",
-	"gleif.org",
-	"moodys.com",
-	"spglobal.com",
-	"fitchratings.com",
-	"morningstarcorp.com",
-	"spcapitaliq.com",
-	"thomsonreuters.com",
-	"factset.com",
-	"bloomberg.com",
-	"ihs.com",
-	"dtcc.com",
-	"federalreserve.gov",
-	"fdic.gov",
-	"occ.treas.gov",
-	"sec.gov",
-	"cftc.gov",
-	"finra.org",
-	"sipc.org",
-	"londonstockexchange.com",
-	"deutsche-boerse.com",
-	"euronext.com",
-	"six-group.com",
-	"tmx.com",
-	"asx.com.au",
-	"hkex.com.hk",
-	"jpx.co.jp",
-	"sgx.com",
-	"krx.co.kr",
-	"bseindia.com",
-	"nseindia.com",
-	"b3.com.br",
-	"moex.com",
-	"saudiexchange.sa",
-	"adx.ae",
-	"dbe.ae",
-	"nasdaqomx.com",
-	"nyse.com",
-	"cboe.com",
-	"cmegroup.com",
-	"theice.com",
-	"eurexchange.com",
-	"lme.com",
-	"sgx.com",
-	"jpx.co.jp",
-	"hkex.com.hk",
-	"asx.com.au",
-	"krx.co.kr",
-	"nse.co.in",
-	"bseindia.com",
-	"b3.com.br",
-	"moex.com",
-	"saudiexchange.sa",
-	"adx.ae",
-	"dgcx.ae",
-	"equinix.com",
-	"digitalrealty.com",
-	"coresite.com",
-	"cyrusone.com",
-	"interxion.com",
-	"telx.com",
-	"telecitygroup.com",
-	"globalswitch.com",
-	"kddi.com",
-	"ntt.com",
-	"tatacommunications.com",
-	"btplc.com",
-	"centurylink.com",
-	"telehouse.net",
-	"verizon.com",
-	"att.com",
-	"level3.com",
-	"orange.com",
-	"telia.com",
-	"telstra.com.au",
-	"hgc.com.hk",
-	"pccwglobal.com",
-	"singtel.com",
-	"globaltransit.net",
-	"chinamobile.com",
-	"chinaunicom.com.cn",
-	"chinatelecom-h.com",
-	"kddi.com",
-	"ntt.com",
-	"tatacommunications.com",
-	"globalservices.bt.com",
-	"centurylink.com",
-	"telehouse.com",
-	"verizonenterprise.com",
-	"business.att.com",
-	"level3.com",
-	"orange-business.com",
-	"teliacarrier.com",
-	"telstraglobal.com",
-	"hgc-intl.com",
-	"pccwglobal.com",
-	"singtel.com",
-	"globaltransit.net",
-	"10086.cn",
-	"chinaunicom.com.hk",
-	"ctexcel.com",
-	"sap.com",
-	"oracle.com",
-	"netapp.com",
-	"vmware.com",
-	"salesforce.com",
-	"ibm.com",
-	"microsoft.com",
-	"google.com",
-	"amazon.com",
-	"adobe.com",
-	"cisco.com",
-	"dell.com",
-	"hpe.com",
-	"hitachivantara.com",
-	"redhat.com",
-	"splunk.com",
-	"servicenow.com",
-	"paloaltonetworks.com",
-	"zscaler.com",
-	"okta.com",
-	"onelogin.com",
-	"pingidentity.com",
-	"sailpoint.com",
-	"cyberark.com",
-	"beyondtrust.com",
-	"centrify.com",
-	"a.root-servers.net",
-	"b.root-servers.net",
-	"c.root-servers.net",
-	"d.root-servers.net",
-	"e.root-servers.net",
-	"f.root-servers.net",
-	"g.root-servers.net",
-	"h.root-servers.net",
-	"i.root-servers.net",
-	"j.root-servers.net",
-	"k.root-servers.net",
-	"l.root-servers.net",
-	"m.root-servers.net",
-	"ns1.icann.org",
-	"ns2.icann.org",
-	"a.iana-servers.net",
-	"b.iana-servers.net",
-	"c.iana-servers.net",
-	"ns1.iana.org",
-	"ns2.iana.org",
-	"ns3.iana.org",
-	"a0.org.afilias-nst.info",
-	"a2.org.afilias-nst.info",
-	"b0.org.afilias-nst.org",
-	"b2.org.afilias-nst.org",
-	"c0.org.afilias-nst.info",
-	"d0.org.afilias-nst.org",
-	"ns1.isc-sns.net",
-	"ns2.isc-sns.com",
-	"ns3.isc-sns.info",
-	"ns1.nic.com",
-	"ns2.nic.com",
-	"a.gtld-servers.net",
-	"b.gtld-servers.net",
-	"c.gtld-servers.net",
-	"d.gtld-servers.net",
-	"e.gtld-servers.net",
-	"f.gtld-servers.net",
-	"g.gtld-servers.net",
-	"h.gtld-servers.net",
-	"i.gtld-servers.net",
-	"j.gtld-servers.net",
-	"k.gtld-servers.net",
-	"l.gtld-servers.net",
-	"m.gtld-servers.net",
-	"a.edu-servers.net",
-	"c.edu-servers.net",
-	"d.edu-servers.net",
-	"f.edu-servers.net",
-	"g.edu-servers.net",
-	"a.mx-servers.net",
-	"b.mx-servers.net",
-	"c.mx-servers.net",
-	"d.mx-servers.net",
-	"e.mx-servers.net",
-	"f.mx-servers.net",
-	"ns1.google.com",
-	"ns2.google.com",
-	"ns3.google.com",
-	"ns4.google.com",
-	"ns1.msft.net",
-	"ns2.msft.net",
-	"ns3.msft.net",
-	"ns4.msft.net",
-	"ns5.msft.net",
-	"ns1.facebook.com",
-	"ns2.facebook.com",
-	"ns3.facebook.com",
-	"ns4.facebook.com",
-	"ns1.amazon.com",
-	"ns2.amazon.com",
-	"ns3.amazon.com",
-	"ns4.amazon.com",
-	"ns1.apple.com",
-	"ns2.apple.com",
-	"ns3.apple.com",
-	"ns4.apple.com",
-	"ns1.cloudflare.com",
-	"ns2.cloudflare.com",
-	"ns3.cloudflare.com",
-	"ns4.cloudflare.com",
-	"ns1.fastly.net",
-	"ns2.fastly.net",
-	"ns3.fastly.net",
-	"ns4.fastly.net",
-	"ns1.akamai.com",
-	"ns2.akamai.com",
-	"ns3.akamai.com",
-	"ns4.akamai.com",
-	"ns1.adnxs.net",
-	"ns2.adnxs.net",
-	"ns3.adnxs.net",
-	"ns4.adnxs.net",
-	"ns1.netflix.com",
-	"ns2.netflix.com",
-	"ns3.netflix.com",
-	"ns4.netflix.com",
-	"ns1-1.akamaitech.net",
-	"ns1-2.akamaitech.net",
-	"ns1-3.akamaitech.net",
-	"ns1-4.akamaitech.net",
-	"ns1-5.akamaitech.net",
-	"ns1-6.akamaitech.net",
-	"ns1-7.akamaitech.net",
-	"ns1-8.akamaitech.net",
-	"ns1-9.akamaitech.net",
-	"ns1-10.akamaitech.net",
-	"ns1-11.akamaitech.net",
-	"ns1-12.akamaitech.net",
-	"ns1-13.akamaitech.net",
-	"ns1-14.akamaitech.net",
-	"ns1-15.akamaitech.net",
-	"ns1-16.akamaitech.net",
-	"ns1-17.akamaitech.net",
-	"ns1-18.akamaitech.net",
-	"ns1-19.akamaitech.net",
-	"ns1-20.akamaitech.net",
-	"ns1.comcast.net",
-	"ns2.comcast.net",
-	"ns3.comcast.net",
-	"ns4.comcast.net",
-	"ns1.verizon.net",
-	"ns2.verizon.net",
-	"ns3.verizon.net",
-	"ns4.verizon.net",
-	"ns1.att.net",
-	"ns2.att.net",
-	"ns3.att.net",
-	"ns4.att.net",
-	"ns1.charter.com",
-	"ns2.charter.com",
-	"ns3.charter.com",
-	"ns4.charter.com",
-	"ns1.cox.net",
-	"ns2.cox.net",
-	"ns3.cox.net",
-	"ns4.cox.net",
-	"ns1.ca-servers.ca",
-	"ns2.ca-servers.ca",
-	"ns3.ca-servers.ca",
-	"ns1.au-servers.au",
-	"ns2.au-servers.au",
-	"ns3.au-servers.au",
-	"ns1.uk-servers.co.uk",
-	"ns2.uk-servers.co.uk",
-	"ns3.uk-servers.co.uk",
-	"ns1.de-servers.de",
-	"ns2.de-servers.de",
-	"ns3.de-servers.de",
-	"ns1.jp-servers.jp",
-	"ns2.jp-servers.jp",
-	"ns3.jp-servers.jp",
-	"ns1.cn-servers.cn",
-	"ns2.cn-servers.cn",
-	"ns3.cn-servers.cn",
-	"ns1.in-servers.in",
-	"ns2.in-servers.in",
-	"ns3.in-servers.in",
-	"ns1.br-servers.br",
-	"ns2.br-servers.br",
-	"ns3.br-servers.br",
-	"ns1.ru-servers.ru",
-	"ns2.ru-servers.ru",
-	"ns3.ru-servers.ru",
-	"ns1.fr-servers.fr",
-	"ns2.fr-servers.fr",
-	"ns3.fr-servers.fr",
-	"ns1.it-servers.it",
-	"ns2.it-servers.it",
-	"ns3.it-servers.it",
-	"ns1.es-servers.es",
-	"ns2.es-servers.es",
-	"ns3.es-servers.es",
-	"ns1.nl-servers.nl",
-	"ns2.nl-servers.nl",
-	"ns3.nl-servers.nl",
-	"ns1.pl-servers.pl",
-	"ns2.pl-servers.pl",
-	"ns3.pl-servers.pl",
-	"ns1.se-servers.se",
-	"ns2.se-servers.se",
-	"ns3.se-servers.se",
-	"ns1.tw-servers.tw",
-	"ns2.tw-servers.tw",
-	"ns3.tw-servers.tw",
-	"ns1.kr-servers.kr",
-	"ns2.kr-servers.kr",
-	"ns3.kr-servers.kr",
-	"ns1.tr-servers.tr",
-	"ns2.tr-servers.tr",
-	"ns3.tr-servers.tr",
-	"ns1.ae-servers.ae",
-	"ns2.ae-servers.ae",
-	"ns3.ae-servers.ae",
-	"ns1.sg-servers.sg",
-	"ns2.sg-servers.sg",
-	"ns3.sg-servers.sg",
-	"ns1.za-servers.za",
-	"ns2.za-servers.za",
-	"ns3.za-servers.za",
-	"windowsupdate.microsoft.com",
-	"update.microsoft.com",
-	"download.windowsupdate.com",
-	"ntservicepack.microsoft.com",
-	"cdn.windowsupdate.com",
-	"statsfe1.ws.microsoft.com",
-	"fe2.update.microsoft.com",
-	"fe3.delivery.mp.microsoft.com",
-	"download.microsoft.com",
-	"tlu.dl.delivery.mp.microsoft.com",
-	"dl.delivery.mp.microsoft.com",
-	"v4.download.windowsupdate.com",
-	"ctldl.windowsupdate.com",
-	"cdn.download.windowsupdate.com",
-	"mesu.apple.com",
-	"appldnld.apple.com",
-	"swcdn.apple.com",
-	"swdist.apple.com",
-	"swdownload.apple.com",
-	"swquery.apple.com",
-	"swscan.apple.com",
-	"pancake.apple.com",
-	"prod.delta.apple.com",
-	"prod.delta.itunes.apple.com",
-	"osxapps.itunes.apple.com",
-	"supportdownload.apple.com",
-	"km.support.apple.com",
-	"gdmf.apple.com",
-	"updates-http.cdn-apple.com",
-	"updates.cdn-apple.com",
-	"gg.google.com",
-	"tools.google.com",
-	"clients2.google.com",
-	"clients4.google.com",
-	"clients1.google.com",
-	"update.googleapis.com",
-	"safebrowsing.googleapis.com",
-	"dl.google.com",
-	"dl-ssl.google.com",
-	"redirector.googlevideo.com",
-	"r1---sn-n4v7sney.googlevideo.com",
-	"r2---sn-n4v7sney.googlevideo.com",
-	"r3---sn-n4v7sney.googlevideo.com",
-	"r4---sn-n4v7sney.googlevideo.com",
-	"r5---sn-n4v7sney.googlevideo.com",
-	"r6---sn-n4v7sney.googlevideo.com",
-	"r7---sn-n4v7sney.googlevideo.com",
-	"r8---sn-n4v7sney.googlevideo.com",
-	"r9---sn-n4v7sney.googlevideo.com",
-	"r10---sn-n4v7sney.googlevideo.com",
-	"r11---sn-n4v7sney.googlevideo.com",
-	"r12---sn-n4v7sney.googlevideo.com",
-	"r13---sn-n4v7sney.googlevideo.com",
-	"r14---sn-n4v7sney.googlevideo.com",
-	"r15---sn-n4v7sney.googlevideo.com",
-	"r16---sn-n4v7sney.googlevideo.com",
-	"r17---sn-n4v7sney.googlevideo.com",
-	"r18---sn-n4v7sney.googlevideo.com",
-	"r19---sn-n4v7sney.googlevideo.com",
-	"r20---sn-n4v7sney.googlevideo.com",
-	"mw1.google.com",
-	"mw2.google.com",
-	"mw3.google.com",
-	"mw4.google.com",
-	"mw5.google.com",
-	"mw6.google.com",
-	"mw7.google.com",
-	"mw8.google.com",
-	"mw9.google.com",
-	"mw10.google.com",
-	"mw11.google.com",
-	"mw12.google.com",
-	"mw13.google.com",
-	"mw14.google.com",
-	"mw15.google.com",
-	"mw16.google.com",
-	"mw17.google.com",
-	"mw18.google.com",
-	"mw19.google.com",
-	"mw20.google.com",
-	"www.download.windowsupdate.com",
-	"dl.windowsupdate.com",
-	"fg.v4.download.windowsupdate.com",
-	"wu.akamaiservice.net",
-	"a248.e.akamai.net",
-	"a104.d.akamai.net",
-	"download.microsoft.com.nsatc.net",
-	"download.windowsupdate.nsatc.net",
-	"wu.msftncsi.com",
-	"sls.update.microsoft.com",
-	"sls.update.microsoft.com.nsatc.net",
-	"swcdn.apple.com.akadns.net",
-	"itunes.apple.com.edgekey.net",
-	"osxapps.itunes.apple.com.edgekey.net",
-	"oscdn.apple.com.akadns.net",
-	"deimos3.apple.com",
-	"albert.apple.com",
-	"gs.apple.com",
-	"ig.apple.com",
-	"gg.apple.com",
-	"c.apple.com",
-	"s.mzstatic.com",
-	"p17-keyvalueservice.icloud.com",
-	"p17-escrowservice.icloud.com",
-	"p17-fmf.icloud.com",
-	"p17-fmip.icloud.com",
-	"p17-ckdatabase.icloud.com",
-	"p17-quota.icloud.com",
-	"configuration.apple.com",
-	"setup.icloud.com",
-	"p17-ckshare.icloud.com",
-	"p17-contactsuploader.icloud.com",
-	"init.ess.apple.com",
-	"p17-iwmb.icloud.com",
-	"p17-docws.icloud.com",
-	"init-p01md.apple.com",
-	"p17-calendars.icloud.com",
-	"appleid.cdn-apple.com",
-	"init.itunes.apple.com",
-	"p17-mbks2.icloud.com",
-	"commondatastorage.googleapis.com",
-	"android.googleapis.com",
-	"android.clients.google.com",
-	"play.googleapis.com",
-	"playatoms-pa.googleapis.com",
-	"android.l.google.com",
-	"mobilesdk-pa.googleapis.com",
-	"firebasestorage.googleapis.com",
-	"firebaseinstallations.googleapis.com",
-	"firebasedynamiclinks.googleapis.com",
-	"crashlytics.googleapis.com",
-	"firebaselogging.googleapis.com",
-	"firebaseinappmessaging.googleapis.com",
-	"firebaseremoteconfig.googleapis.com",
-	"people-pa.googleapis.com",
-	"firebaseperformance.googleapis.com",
-	"scone-pa.googleapis.com",
-	"firebaseappcheck.googleapis.com",
-	"fcmregistrations.googleapis.com",
-	"mdh-pa.googleapis.com",
-	"growth-pa.googleapis.com",
-	"instantmessaging-pa.googleapis.com",
-	"tpc.googlesyndication.com",
-	"www-googletagmanager.l.google.com",
-	"ssl-google-analytics.l.google.com",
-	"www-google-analytics.l.google.com",
-	"www.googletagservices.com",
-	"www.google-analytics.com",
-	"firebasepubsub-pa-clients6a.googleapis.com",
-	"fonts.googleapis.com",
-	"firebaserules.googleapis.com",
-	"mobilecrashreporting.googleapis.com",
-	"dns.google",
-	"connectivitycheck.gstatic.com",
-	"googleapis.l.google.com",
-	"update.crashlytics.com",
-	"incoming.telemetry.mozilla.org",
-	"redirector.gvt1.com",
-	"apis.google.com",
-	"people-pa.clients6.google.com",
-	"surveys.google.com",
-	"clientservices.googleapis.com",
-	"plus.google.com",
-	"adservice.google.com",
-	"video-stats.video.google.com",
-	"googleads.g.doubleclick.net",
-	"firebase.google.com",
-	"firebaseremoteconfig.clients6.google.com",
-	"accounts.youtube.com",
-	"oauthaccountmanager.googleapis.com",
-	"update.googleapis.com",
-	"safebrowsing.clients.google.com",
-	"app-measurement.com",
-	"mobilesdk-pa.clients6.google.com",
-	"instantmessaging-pa.clients6.google.com",
-	"update.mgmt.google.com",
-	"alt9-mtalk.google.com",
-	"alt8-mtalk.google.com",
-	"alt7-mtalk.google.com",
-	"alt6-mtalk.google.com",
-	"alt5-mtalk.google.com",
-	"alt4-mtalk.google.com",
-	"alt3-mtalk.google.com",
-	"alt2-mtalk.google.com",
-	"alt1-mtalk.google.com",
-	"www3.l.google.com",
-	"www4.l.google.com",
-	"fcmregistrations.googleapis.clients.google.com",
-	"peopleapi.clients6.google.com",
-	"identitytoolkit.googleapis.com",
-	"securetoken.googleapis.com",
-	"firestore.googleapis.com",
-	"pagead2.googlesyndication.com",
-	"adservice.google.ca",
-	"adservice.google.fr",
-	"adservice.google.com.hk",
-	"adservice.google.de",
-	"adservice.google.co.uk",
-	"adservice.google.com.au",
-	"adservice.google.co.in",
-	"www.uscis.gov",
-	"www.gov.uk",
-	"www.canada.ca",
-	"www.border.gov.au",
-	"www.vfsglobal.com",
-	"www.immigration.govt.nz",
-	"www.inis.gov.ie",
-	"www.migrationsverket.se",
-	"www.bamf.de",
-	"www.diplomatie.gouv.fr",
-	"www.gov.il",
-	"www.udi.no",
-	"www.migri.fi",
-	"www.nyidanmark.dk",
-	"www.sem.admin.ch",
-	"www.rvv.government.bg",
-	"www.mzv.cz",
-	"www.vie.asz.hr",
-	"www.raiiministeerium.ee",
-	"www.bmbah.hu",
-	"www.pmlp.gov.lv",
-	"www.migracija.lt",
-	"www.servizz.gov.mt",
-	"www.gov.pl",
-	"www.imigrante.sef.pt",
-	"www.mae.ro",
-	"www.minv.sk",
-	"www.gov.si",
-	"www.interior.gob.es",
-	"www.government.se",
-	"www.ukvisas.gov.uk",
-	"www.mfa.gov.il",
-	"www.cgitoronto.gov.in",
-	"www.hongkongvisa.gov.hk",
-	"www.ind.nl",
-	"www.mofa.go.jp",
-	"www.visa.go.kr",
-	"www.evisa.gov.kw",
-	"www.evisatraveller.mfa.gov.qa",
-	"www.mofa.gov.sa",
-	"www.ica.gov.ae",
-	"www.immigration.go.th",
-	"www.mfa.gov.sg",
-	"www.vfsglobal.com",
-	"www.imi.gov.my",
-	"www.pdi.gov.ph",
-	"www.kbri-canberra.org.au",
-	"www.visa.mofa.gov.vn",
-	"www.indianvisaonline.gov.in",
-	"www.boca.gov.tw",
-	"www.immigration.gov.hk",
-	"www.ica.gov.sg",
-	"www.eta.gov.lk",
-	"www.mip.gov.ua",
-	"www.evisa.gov.az",
-	"www.mfa.gov.by",
-	"www.geoconsul.gov.ge",
-	"www.gov.kz",
-	"www.evisa.gov.kg",
-	"www.mfa.gov.md",
-	"www.evisa.tj",
-	"www.turkmenistan.gov.tm",
-	"www.mfa.uz",
-	"www.evisa.gov.af",
-	"www.visa.gov.bd",
-	"www.mfa.gov.bt",
-	"www.online.nepalimmigration.gov.np",
-	"www.immigration.gov.pk",
-	"www.migrations.gov.mg",
-	"www.visa.gov.ci",
-	"www.visa.gov.gn",
-	"www.ecitizen.go.ke",
-	"www.immigration.gov.ng",
-	"www.migration.gov.rw",
-	"www.immigration.go.ug",
-	"www.immigration.go.tz",
-	"www.moha.gov.zm",
-	"www.evisa.gov.bw",
-	"www.immigration.gov.mw",
-	"www.smi.gov.sc",
-	"www.emigration.gov.za",
-	"www.nacionalidade.gov.cv",
-	"www.consuladovirtual.mirex.gob.do",
-	"www.gis.gov.gy",
-	"www.migracion.gob.pa",
-	"www.serviciosmigratorios.gob.do",
-	"www.cancilleria.gob.ar",
-	"www.portaldeimigracao.mj.gov.br",
-	"www.tramites.minrel.gob.cl",
-	"www.migracioncolombia.gov.co",
-	"www.gob.ec",
-	"www.vtv.migracion.gob.bo",
-	"www.mre.gov.py",
-	"www.migraciones.gub.uy",
-	"www.bmi.gv.at",
-	"www.visa.bmeia.gv.at",
-	"www.dofi.ibz.be",
-	"www.cyprus.gov.cy",
-	"www.mzv.cz",
-	"www.nyidanmark.dk",
-	"www.vm.ee",
-	"www.formin.finland.fi",
-	"www.diplomatie.gouv.fr",
-	"www.auswaertiges-amt.de",
-	"www.mfa.gr",
-	"www.konzuliszolgalat.kormany.hu",
-	"www.utl.is",
-	"www.vistoperitalia.esteri.it",
-	"www.am.gov.lv",
-	"www.urm.lt",
-	"www.guichet.public.lu",
-	"www.identitymalta.com",
-	"www.udi.no",
-	"www.gov.pl",
-	"www.vistos.mne.pt",
-	"www.mae.ro",
-	"www.mzv.sk",
-	"www.mzz.gov.si",
-	"www.exteriores.gob.es",
-	"www.government.se",
-	"www.eda.admin.ch",
-	"www.gov.uk",
-	"www.gov.uk",
-	"www.migri.fi",
-	"www.am.gov.lv",
-	"www.mae.ro",
-	"www.inis.gov.ie",
-	"www.formin.finland.fi",
-	"www.migrationsverket.se",
-	"www.diplomatie.gouv.fr",
-	"www.governo.it",
-	"www.eda.admin.ch",
-	"www.cyprus.gov.cy",
-	"www.india.blsspainvisa.com",
-	"www.india.blsgermanyvisa.com",
-	"www.vfsglobal.ca",
-	"www.jpa-visaservices.com",
-	"www.uae-embassy.ae",
-	"www.vfsglobal.com",
-	"www.evisa.gov.tr",
-	"www.musavirlikler.gov.tr",
-	"www.gdrfad.gov.ae",
-	"www.mofa.gov.sa",
-	"www.visatohungary.gov.hu",
-	"www.evisa.mfa.ir",
-	"www.kuwaitvisa.com",
-	"www.visa.mofa.gov.om",
-	"www.moi.gov.qa",
-	"www.mofaic.gov.ae",
-	"www.mfa.gov.il",
-	"www.cis.gov.jo",
-	"www.mfa.gov.lb",
-	"www.mfa.gov.sy",
-	"www.visa.visitsaudi.com",
-	"www.visasturkey.com",
-	"www.egypt.travel",
-	"www.algeria.com",
-	"www.diplomatie.gov.dz",
-	"www.visa.gov.ma",
-	"www.mfa.gov.tn",
-	"www.dz.embassyinformation.com",
-	"www.ly.embassyinformation.com",
-	"www.sd.embassyinformation.com",
-	"www.ss.embassyinformation.com",
-	"www.interior.pna.ps",
-	"www.evisa.gov.bh",
-	"www.evisa.rop.gov.om",
-	"www.mfa.gov.tm",
-	"www.caa.gov.az",
-	"www.evisa.mfa.tj",
-	"www.evisa.e-gov.az",
-	"www.evisa.gov.kz",
-	"www.evisa.gov.uz",
-	"www.am.gov.lv",
-	"www.mfa.gov.md",
-	"www.turkmenistanembassy.org",
-	"www.mfa.tj",
-	"www.evisa.immigration.gov.ng",
-	"www.immigration.go.ke",
-	"www.migration.gov.rw",
-	"www.immigration.go.ug",
-	"www.immigration.go.tz",
-	"www.zambiaimmigration.gov.zm",
-	"www.evisa.gov.et",
-	"www.evisa.gov.na",
-	"www.visas.gov.mz",
-	"www.botswanaembassy.org",
-	"www.dha.gov.za",
-	"www.gov.mu",
-	"www.gov.mu",
-	"www.gov.mu",
-	"www.gov.mu",
-	"www.gov.sc",
-	"www.egov.sc",
-	"www.egov.sc",
-	"www.migration.gov.mg",
-	"www.visa.gouv.ci",
-	"www.immigration.go.zm",
-	"www.mha.gov.gi",
-	"www.dgie.ci",
-	"api.salesforce.com",
-	"graph.microsoft.com",
-	"api.amazonwebservices.com",
-	"api.box.com",
-	"api.dropboxapi.com",
-	"api.oracle.com",
-	"api.sap.com",
-	"api.zoom.us",
-	"api.docusign.net",
-	"api.hubspot.com",
-	"api.atlassian.com",
-	"api.zendesk.com",
-	"api.intuit.com",
-	"api.slack.com",
-	"api.twilio.com",
-	"api.stripe.com",
-	"api.paypal.com",
-	"api.authorize.net",
-	"api.square.com",
-	"api.braintreegateway.com",
-	"api.freshworks.com",
-	"api.shopify.com",
-	"api.bigcommerce.com",
-	"api.magento.com",
-	"api.woocommerce.com",
-	"api.squarespace.com",
-	"api.wix.com",
-	"api.godaddy.com",
-	"api.mongodb.com",
-	"api.segment.com",
-	"api.mixpanel.com",
-	"api.optimizely.com",
-	"api.pendo.io",
-	"api.fullstory.com",
-	"api.heap.io",
-	"api.amplitude.com",
-	"api.appcues.com",
-	"api.intercom.io",
-	"api.drift.com",
-	"api.salesforceliveagent.com",
-	"api.livechat.com",
-	"api.olark.com",
-	"api.tawk.to",
-	"api.purechat.com",
-	"api.snapengage.com",
-	"api.liveperson.net",
-	"api.boldchat.com",
-	"api.comm100.com",
-	"api.kayako.com",
-	"api.freshdesk.com",
-	"api.desk.com",
-	"api.zendesk.com",
-	"api.uservoice.com",
-	"api.getsatisfaction.com",
-	"api.helpscout.net",
-	"api.elevio.com",
-	"api.helpjuice.com",
-	"api.groovehq.com",
-	"api.supportbee.com",
-	"api.enchant.com",
-	"api.frontapp.com",
-	"api.helpshift.com",
-	"api.happyfox.com",
-	"api.zohodesk.com",
-	"api.freshservice.com",
-	"api.jitbit.com",
-	"api.osticket.com",
-	"api.reamaze.com",
-	"api.teamwork.com",
-	"api.ladesk.com",
-	"api.deskpro.com",
-	"api.mhelpdesk.com",
-	"api.usabilityhub.com",
-	"api.crazyegg.com",
-	"api.hotjar.com",
-	"api.mouseflow.com",
-	"api.luckyorange.com",
-	"api.inspectlet.com",
-	"api.fullstory.com",
-	"api.quantummetric.com",
-	"api.contentsquare.com",
-	"api.decibelinsight.com",
-	"api.logrocket.com",
-	"api.smartlook.com",
-	"api.clarity.ms",
-	"api.useriq.com",
-	"api.pendo.io",
-	"api.walkme.com",
-	"api.appcues.com",
-	"api.userpilot.com",
-	"api.chameleon.io",
-	"api.userlane.com",
-	"api.whatfix.com",
-	"api.nickelled.com",
-	"api.inline-manual.com",
-	"api.helppier.com",
-	"api.iorad.com",
-	"api.toonimo.com",
-	"api.iridize.com",
-	"api.userlevel.com",
-	"api.userlane.com",
-	"api.salesforceiq.com",
-	"api.base.com",
-	"api.getbase.com",
-	"api.close.com",
-	"api.copper.com",
-	"api.pipedrive.com",
-	"api.pipelinedeals.com",
-	"api.prosperworks.com",
-	"api.capsulecrm.com",
-	"api.nutshell.com",
-	"api.getresponse.com",
-	"api.activecampaign.com",
-	"api.constantcontact.com",
-	"api.mailchimp.com",
-	"api.sendgrid.com",
-	"api.aweber.com",
-	"api.mailgun.net",
-	"api.sparkpost.com",
-	"api.litmus.com",
-	"api.customer.io",
-	"api.bronto.com",
-	"api.drip.com",
-	"api.klaviyo.com",
-	"api.omnisend.com",
-	"api.getnotify.com",
-	"api.sendlane.com",
-	"api.revue.co",
-	"api.convertkit.com",
-	"api.salesmanago.com",
-	"api.getresponse360.com",
-	"api.woodpecker.co",
-	"api.lemlist.com",
-	"api.reply.io",
-	"api.yesware.com",
-	"api.outreach.io",
-	"api.salesloft.com",
-	"api.cirrusinsight.com",
-	"api.mixmax.com",
-	"api.toutapp.com",
-	"api.gong.io",
-	"api.chorus.ai",
-	"api.wingman.ai",
-	"api.exceed.ai",
-	"api.dialpad.com",
-	"api.ujet.co",
-	"api.talkdesk.com",
-	"api.ringcentral.com",
-	"api.niceincontact.com",
-	"api.five9.com",
-	"api.genesys.com",
-	"api.mitel.com",
-	"api.8x8.com",
-	"api.nextiva.com",
-	"api.vonage.com",
-	"api.twilio.com",
-	"api.plivo.com",
-	"api.bandwidth.com",
-	"api.signalwire.com",
-	"api.telnyxsms.com",
-	"api.messagebird.com",
-	"api.nexmo.com",
-	"api.sinch.com",
-	"api.kaleyra.com",
-	"api.clicksend.com",
-	"api.cm.com",
-	"api.messente.com",
-	"api.infobip.com",
-	"api.46elks.com",
-	"api.esendex.com",
-	"api.textlocal.com",
-	"api.smsapi.com",
-	"api.twizo.com",
-	"api.burstsms.com",
-	"api.pswin.com",
-	"api.mobizon.com",
-	"api.smsglobal.com",
-	"api.nexusthemes.com",
-	"api.thesmsworks.co.uk",
-	"api.textanywhere.net",
-	"api.textmarketer.co.uk",
-	"api.firetext.co.uk",
-	"api.fastsms.co.uk",
-	"api.mobivate.com",
-	"api.whispir.com",
-	"api.messagebird.com",
-	"api.plivosms.com",
-	"api.bulksms.com",
-	"api.zensend.com",
-	"api.securesmsc.com",
-	"api.clicksend.com",
-	"api.textmagic.com",
-	"api.smsbroadcast.com.au",
-	"api.directsms.com.au",
-	"api.smscentral.com.au",
-	"api.smsglobal.com",
-	"api.smsbroadcast.co.uk",
-	"api.clockworksms.com",
-	"api.messagenet.com.au",
-	"api.smsdash.com",
-	"time.nist.gov",
-	"time.google.com",
-	"ntp.org",
-	"pool.ntp.org",
-	"time.windows.com",
-	"time.euro.apple.com",
-	"time.apple.com",
-	"ntp.ubuntu.com",
-	"time.cloudflare.com",
-	"ntp.amazon.com",
-	"time.facebook.com",
-	"ntp1.google.com",
-	"ntp2.google.com",
-	"ntp3.google.com",
-	"ntp4.google.com",
-	"ntp.stanford.edu",
-	"ntp.nasa.gov",
-	"ntp1.nasa.gov",
-	"ntp2.nasa.gov",
-	"ntp.nist.gov",
-	"nist1.nist.gov",
-	"nist2.nist.gov",
-	"nist3.nist.gov",
-	"ntp1.oracle.com",
-	"ntp2.oracle.com",
-	"ntp3.oracle.com",
-	"ntp-s1.cais.rnp.br",
-	"a.st1.ntp.br",
-	"b.st1.ntp.br",
-	"c.st1.ntp.br",
-	"d.st1.ntp.br",
-	"ntp.shoa.cl",
-	"ntp.shoa.cl.backup",
-	"ntp0.nl.net",
-	"ntp1.nl.net",
-	"ntp2.nl.net",
-	"ntp.myfloridacity.us",
-	"gnomon.cc.columbia.edu",
-	"navobs1.wustl.edu",
-	"tick.ucla.edu",
-	"tock.ucla.edu",
-	"ntp1.paris-sciences.fr",
-	"ntp2.paris-sciences.fr",
-	"ntp3.paris-sciences.fr",
-	"ntp0.fau.de",
-	"ntp1.fau.de",
-	"ntp2.fau.de",
-	"ntp3.fau.de",
-	"ptbtime1.ptb.de",
-	"ptbtime2.ptb.de",
-	"ptbtime3.ptb.de",
-	"time1.isu.net.sa",
-	"time2.isu.net.sa",
-	"ntp.metas.ch",
-	"ntp.switch.ch",
-	"ntp0.ethz.ch",
-	"ntp1.ethz.ch",
-	"ntp2.ethz.ch",
-	"ntp3.ethz.ch",
-	"ntp0.arnes.si",
-	"ntp1.arnes.si",
-	"ntp2.arnes.si",
-	"ntp-1.oma.be",
-	"ntp-2.oma.be",
-	"ntp.belnet.be",
-	"ntp.nic.cz",
-	"ntp.cesnet.cz",
-	"ntp.fel.cvut.cz",
-	"ntp1.nic.cz",
-	"ntp2.nic.cz",
-	"ntp3.nic.cz",
-	"ntp4.nic.cz",
-	"ntp5.nic.cz",
-	"ntp.telenet.be",
-	"ntp.belnet.be",
-	"ntp.myfloridacity.us",
-	"ntp.myfloridacity.us",
-	"ntp1.vniiftri.ru",
-	"ntp2.vniiftri.ru",
-	"ntp.ix.ru",
-	"ntp1.ix.ru",
-	"ntp2.ix.ru",
-	"ntp3.ix.ru",
-	"ntp4.ix.ru",
-	"ntp21.vniiftri.ru",
-	"ntp1.niiftri.irkutsk.ru",
-	"ntp2.niiftri.irkutsk.ru",
-	"ntp1.ntp-servers.net",
-	"ntp2.ntp-servers.net",
-	"ntp3.ntp-servers.net",
-	"ntp4.ntp-servers.net",
-	"ntp5.ntp-servers.net",
-	"ntp6.ntp-servers.net",
-	"ntp.remco.org",
-	"ntp0.nl.uu.net",
-	"ntp1.nl.uu.net",
-	"ntp2.nl.uu.net",
-	"ntp0.twente.nl",
-	"ntp1.twente.nl",
-	"ntp.online.net",
-	"ntp.eqx.net",
-	"ntp.nrc.ca",
-	"ntp1.nrc.ca",
-	"ntp2.nrc.ca",
-	"ntp3.nrc.ca",
-	"ntp1.net.berkeley.edu",
-	"ntp2.net.berkeley.edu",
-	"ntp1.cmc.ec.gc.ca",
-	"time.chu.nrc.ca",
-	"time-a.timefreq.bldrdoc.gov",
-	"time-b.timefreq.bldrdoc.gov",
-	"time-c.timefreq.bldrdoc.gov",
-	"time.nrc.ca",
-	"chime.science.ubc.ca",
-	"tick.ucd.ie",
-	"tock.ucd.ie",
-	"ntp.adelaide.edu.au",
-	"ntp.cs.mu.oz.au",
-	"augean.eleceng.adelaide.edu.au",
-	"ntp.cs.usyd.edu.au",
-	"ntp.cs.umn.edu",
-	"ntp.maths.tcd.ie",
-	"ntp1.ntp.org.il",
-	"ntp2.ntp.org.il",
-	"ntp3.ntp.org.il",
-	"ntp.ring.gr.jp",
-	"ntp-1.cin.ufpe.br",
-	"ntp-2.cin.ufpe.br",
-	"ntp.cais.rnp.br",
-	"ntp.ime.usp.br",
-	"ntp1.ptb.de",
-	"ntp2.ptb.de",
-	"ntp3.ptb.de",
-	"ntp.oma.be",
-	"ntp.on.net",
-	"ntp.xs4all.nl",
-	"ntp1.ntp.org.cn",
-	"ntp2.ntp.org.cn",
-	"ntp.lth.se",
-	"ntp2.lth.se",
-	"timehost.lysator.liu.se",
-	"ntp1.sp.se",
-	"ntp2.sp.se",
-	"rustime01.rus.uni-stuttgart.de",
-	"rustime02.rus.uni-stuttgart.de",
-	"ntp1.gbg.netnod.se",
-	"ntp2.gbg.netnod.se",
-	"ntp1.mmo.netnod.se",
-	"ntp2.mmo.netnod.se",
-	"ntp1.sth.netnod.se",
-	"ntp2.sth.netnod.se",
-	"ntp1.sth.seacomnet.se",
-	"ntp2.sth.seacomnet.se",
-	"ntp1.csir.co.za",
-	"ntp2.csir.co.za",
-	"clepsydra.dec.com",
-	"clock-1.cs.cmu.edu",
-	"clock-2.cs.cmu.edu",
-	"ntp0.cornell.edu",
-	"ntp1.cornell.edu",
-	"ntp-1.ece.cmu.edu",
-	"ntp-2.ece.cmu.edu",
-	"sundial.columbia.edu",
-	"aurora.alaska.edu",
-	"ntp1.creativemedia.net",
-	"ntp2.creativemedia.net",
-	"ntp1.waia.asn.au",
-	"ntp2.waia.asn.au",
-	"gps.layer42.net",
-	"rdg-gps-1.net.rdg.ac.uk",
-	"rdg-gps-2.net.rdg.ac.uk",
-	"ics-gps-1.net.ics.uci.edu",
-	"usno.afit.edu",
-	"usno.dejablue.com",
-	"usno.palna.com",
-	"usno.qsinet.com",
-	"usno.airnav.com",
-	"hpb.nist.gov",
-	"ats1.nist.gov",
-	"nist1-dc.glassey.com",
-	"nist1-sj.glassey.com",
-	"nist1-macon.glassey.com",
-	"usno.pa-x.dec.com",
-	"clock.psu.edu",
-	"ntp1.umb.edu",
-	"ntp1.sjsu.edu",
-	"ntp3.ntp.org",
-	"ntp.ucsd.edu",
-	"timekeeper.isi.edu",
-	"time_a.nist.gov",
-	"time_b.nist.gov",
-	"time_nw.nist.gov",
-	"ntp1.glbn.nist.gov",
-	"ntp2.glbn.nist.gov",
-	"ntp-s1.cais.rnp.br",
-	"a.st1.ntp.br",
-	"b.st1.ntp.br",
-	"c.st1.ntp.br",
-	"d.st1.ntp.br",
-	"s1a.time.edu.cn",
-	"swcdn.apple.com",
-	"swdist.apple.com",
-	"swdownload.apple.com",
-	"swquery.apple.com",
-	"swscan.apple.com",
-	"itunes.apple.com",
-	"appldnld.apple.com",
-	"configurator.apple.com",
-	"gnf-mdn.apple.com",
-	"gnf-mr.apple.com",
-	"ig.apple.com",
-	"mesu.apple.com",
-	"ns.itunes.apple.com",
-	"oscdn.apple.com",
-	"osrecovery.apple.com",
-	"supportdownload.apple.com",
-	"sw-update.apple.com",
-	"updates-http.cdn-apple.com",
-	"updates.cdn-apple.com",
-	"xp.apple.com",
-	"albert.apple.com",
-	"gs.apple.com",
-	"ipod.apple.com",
-	"itunesu.assets.itunes.com",
-	"su.itunes.apple.com",
-	"updates-secure.apple.com",
-	"updates-secure-cn.apple.com",
-	"updates-staging.apple.com",
-	"updates.apple.com",
-	"valid.apple.com",
-	"aaplimg.com",
-	"akadns.net",
-	"akamai.net",
-	"akamaiedge.net",
-	"deploy.apple.com",
-	"deploy.static.akamaitechnologies.com",
-	"icloud.cdn-apple.com",
-	"images.apple.com",
-	"init-p01md.apple.com",
-	"init-p01st.push.apple.com",
-	"init.ess.apple.com",
-	"init.gc.apple.com",
-	"init.itunes.apple.com",
-	"iosapps.itunes.apple.com",
-	"is1.mzstatic.com",
-	"is2.mzstatic.com",
-	"is3.mzstatic.com",
-	"is4.mzstatic.com",
-	"is5.mzstatic.com",
-	"itunes-apple.com.akadns.net",
-	"itunes.apple.com.akadns.net",
-	"itunes.com",
-	"itunesconnect.apple.com",
-	"metrics.apple.com",
-	"osxapps.itunes.apple.com",
-	"pad.itunes.apple.com",
-	"sandbox.itunes.apple.com",
-	"play.itunes.apple.com",
-	"se.itunes.apple.com",
-	"sp.itunes.apple.com",
-	"streamingaudio.itunes.apple.com",
-	"su-cdn.itunes-apple.com.akadns.net",
-	"upp.itunes.apple.com",
-	"valid.ess.apple.com",
-	"www.apple.com",
-	"www.icloud.com",
-	"config.apple.com",
-	"crl.apple.com",
-	"crl.entrust.net",
-	"crl3.digicert.com",
-	"crl4.digicert.com",
-	"developer.apple.com",
-	"devimages.apple.com",
-	"devstreaming.apple.com",
-	"doh.dns.apple.com",
-	"download.developer.apple.com",
-	"gs-loc.apple.com",
-	"gs-loc-cn.apple.com",
-	"gsp10-ssl.apple.com",
-	"gsp4-cn-ssl.apple.com",
-	"gsp4-ssl.apple.com",
-	"gsp5-ssl.apple.com",
-	"gsp64-ssl.apple.com",
-	"gsp64-ssl-cn.apple.com",
-	"gspe1-ssl.apple.com",
-	"gspe11-ssl.apple.com",
-	"gspe12-ssl.apple.com",
-	"gspe19-ssl.apple.com",
-	"gspe2-ssl.apple.com",
-	"gspe21-ssl.apple.com",
-	"gspe35-ssl.apple.com",
-	"gspe4-ssl.ls.apple.com",
-	"gspe4-ssl.apple.com",
-	"gspe5-ssl.apple.com",
-	"gspe85-ssl.apple.com",
-	"humb.apple.com",
-	"iadsdk.apple.com",
-	"icloud-content.com",
-	"images.apple.com",
-	"init-p01st.apple.com",
-	"init.push.apple.com",
-	"iphonesubmissions.apple.com",
-	"ocsp.apple.com",
-	"ocsp.entrust.net",
-	"ocsp.verisign.net",
-	"pancake.apple.com",
-	"phobos.apple.com",
-	"prod-support.apple-support.akadns.net",
-	"setup.fe.apple-dns.net",
-	"setup.icloud.com",
-	"store.apple.com",
-	"support.apple.com",
-	"swscan.apple.com.edgekey.net",
-	"time-ios.apple.com",
-	"time-ios.g.aaplimg.com",
-	"time-macos.apple.com",
-	"time.apple.com",
-	"time.asia.apple.com",
-	"time.euro.apple.com",
-	"updates-http.cdn-apple.com",
-	"updates-http.cdn-apple.com.akadns.net",
-	"updates.cdn-apple.com",
-	"valid.apple.com.edgekey.net",
-	"valid.origin-apple.com.akadns.net",
-	"www.icloud.com.akadns.net",
-	"aod.itunes.apple.com",
-	"cdn-cn1.apple-mapkit.com",
-	"cdn-cn2.apple-mapkit.com",
-	"cdn-cn3.apple-mapkit.com",
-	"cdn-cn4.apple-mapkit.com",
-	"cdn1.apple-mapkit.com",
-	"cdn2.apple-mapkit.com",
-	"cdn3.apple-mapkit.com",
-	"cdn4.apple-mapkit.com",
-	"cds-cdn.v.aaplimg.com",
-	"cds.apple.com",
-	"deimos3.apple.com",
-	"itsliveradio.apple.com",
-	"mesu-cdn.apple.com.akadns.net",
-	"mesu-china.apple.com.akadns.net",
-	"mesu.apple.com.akadns.net",
-	"itunes.apple.com.edgekey.net",
-	"music.apple.com",
-	"osxapps.itunes.apple.com.edgekey.net",
-	"radio-services.itunes.apple.com",
-	"setup.icloud.com.akadns.net",
-	"stocks-sparkline-cache.apple.com",
-	"store.apple.com.edgekey.net",
-	"store.storeimages.apple.com.akadns.net",
-	"upp.itunes.apple.com.edgekey.net",
-	"uts-api.itunes.apple.com",
-	"www.apple.com.edgekey.net",
-	"xp.apple.com.akadns.net",
-	"audiobooks.itunes.apple.com",
-	"cdn.apple-mapkit.com",
-	"cdn1.apple-mapkit.com",
-	"cdn2.apple-mapkit.com",
-	"cdn3.apple-mapkit.com",
-	"cdn4.apple-mapkit.com",
-	"cma.itunes.apple.com",
-	"communities.apple.com",
-	"cs.apple.com",
-	"cstat.apple.com",
-	"cstat.cdn-apple.com",
-	"discussions.apple.com",
-	"downloads.apple.com",
-	"feedbackws.icloud.com",
-	"fides-pol.apple.com",
-	"gs-loc-cn.ls-apple.com.akadns.net",
-	"gsa.apple.com",
-	"gsas.apple.com",
-	"gspe35-ssl-cn.ls.apple.com",
-	"icloud-cdn.icloud.com",
-	"idmsa.apple.com",
-	"iforgot.apple.com",
-	"imeviewer-st11.apple.com",
-	"iphonesupportfiles.apple.com",
-	"itunes.mzstatic.com",
-	"itunesu.assets.itunes.com",
-	"itunesu.assets.itunes.com.akadns.net",
-	"itunesu.itunes.apple.com",
-	"km.support.apple.com",
-	"ls.apple.com",
-	"movies.apple.com",
-	"oscdn.apple.com.akadns.net",
-	"osxapps.itunes.apple.com.edgesuite.net",
-	"osxsupportfiles.apple.com",
-	"ppq.apple.com",
-	"publicassets.cdn-apple.com",
-	"reportingitc.apple.com",
-	"s.mzstatic.com",
-	"selfsolve.apple.com",
-	"ssl-cstat.cdn-apple.com",
-	"stocks-cdn.apple.com",
-	"stocks-cdn.ls.apple.com",
-	"streamingaudio.itunes.apple.com.edgekey.net",
-	"support.ess.apple.com",
-	"swdist.apple.com.akadns.net",
-	"syncwow.apple.com",
-	"tv.apple.com",
-	"update.hikvision.com",
-	"update.dahuasecurity.com",
-	"update.tp-link.com.cn",
-	"update.yealink.com",
-	"update.uniview.com",
-	"update.huawei.com",
-	"update.zte.com.cn",
-	"update.h3c.com",
-	"update.ruijie.com.cn",
-	"update.maipu.com",
-	"update.tenda.com.cn",
-	"update.mercurycom.com.cn",
-	"update.dbappsecurity.com.cn",
-	"update.baiducloud.com",
-	"update.alibabacloud.com",
-	"update.jd.com",
-	"update.lenovo.com.cn",
-	"update.mi.com",
-	"update.oppo.com",
-	"update.vivo.com.cn",
-	"update.oneplus.cn",
-	"update.meizu.com",
-	"update.gionee.com",
-	"update.smartisan.com",
-	"update.nubia.com",
-	"update.lephone.com",
-	"update.360.cn",
-	"update.tclcom.com",
-	"update.hisense.com",
-	"update.skyworth.com",
-	"update.changhong.com",
-	"update.konka.com",
-	"update.haier.com",
-	"update.zte.com.cn",
-	"update.fiberhome.com",
-	"update.seuic.com",
-	"update.sunmi.com",
-	"update.szzt.com",
-	"update.xiaongmai.com",
-	"update.hankvision.com",
-	"update.taichuan.com",
-	"update.gospell.com",
-	"update.sdmctech.com",
-	"update.hanbang.com",
-	"update.kedacom.com",
-	"update.wanbangde.com",
-	"update.dualshine.com",
-	"update.jovision.com",
-	"update.topsee.com",
-	"update.bolin.com.cn",
-	"update.infinova.com",
-	"update.jiaodajinggong.com",
-	"update.intl-intel.com",
-	"update.amd.com.cn",
-	"update.nvidia.cn",
-	"update.msi.com",
-	"update.asus.com.cn",
-	"update.gigabyte.cn",
-	"update.colorful.cn",
-	"update.galaxy.com",
-	"update.zotac.com",
-	"update.xfx.com",
-	"update.his.com",
-	"update.sapphiretech.com",
-	"update.powercolor.com",
-	"update.yeston.net",
-	"update.maxsun.com.cn",
-	"update.soyo.com",
-	"update.leadtek.com.tw",
-	"update.afox.com.tw",
-	"update.andyson.com",
-	"update.biostar.com.tw",
-	"update.ecs.com.tw",
-	"update.jetway.com.tw",
-	"update.dfi.com",
-	"update.aopen.com",
-	"update.ibase.com.tw",
-	"update.sparkle.com.tw",
-	"update.asrock.com",
-	"update.supermicro.com",
-	"update.foxconn.com",
-	"update.shuttle.com",
-	"update.acer.com",
-	"update.dell.com.cn",
-	"update.hp.com.cn",
-	"update.benq.com",
-	"update.nec-display.com",
-	"update.philips.com.cn",
-	"update.aoc.com",
-	"update.rapoo.com",
-	"update.a4tech.com",
-	"update.logitech.com.cn",
-	"update.microsoft.com.cn",
-	"update.cherry.de",
-	"update.corsair.com",
-	"update.razer.com",
-	"update.steelseries.com",
-	"update.roccat.org",
-	"update.madcatz.com",
-	"update.thermaltake.com",
-	"update.dlink.com",
-	"update.netgear.com.cn",
-	"update.trendnet.com",
-	"update.zyxel.com.cn",
-	"update.cisco.com",
-	"update.linksys.com",
-	"update.buffalo-asia.com",
-	"update.qnap.com",
-	"update.synology.com",
-	"update.asustor.com",
-	"update.drobo.com",
-	"update.wdc.com",
-	"update.seagate.com",
-	"update.ironwolf.com",
-	"update.hgst.com",
-	"update.lacie.com",
-	"update.g-technology.com",
-	"update.clevoad.com",
-	"update.liteon.com",
-	"update.plextor.com",
-	"update.kingston.com",
-	"update.sandisk.com.cn",
-	"update.westerndigital.com",
-	"update.samsung.com",
-	"update.crucial.com",
-	"update.micron.com",
-	"update.adata.com",
-	"update.apacer.com",
-	"update.transcend-info.com",
-	"update.sony.com.cn",
-	"update.canon.com.cn",
-	"update.epson.com.cn",
-	"update.brother.com.cn",
-	"update.hp.com",
-	"update.lexmark.com",
-	"update.xerox.com",
-	"update.pantum.com",
-	"update.founder.com.cn",
-	"update.lenovo.com.cn",
-	"update.nec-printing.com",
-	"update.oki.com",
-	"update.kyocera.com",
-	"update.sharp-world.com",
-	"update.toshiba-tec.com.cn",
-	"update.kodak.com",
-	"update.ricoh.com",
-	"update.fujifilm.com",
-	"update.konicaminolta.com",
-	"update.samsung.com",
-	"update.sindoh.com",
-	"update.Dell.com",
-	"update.viewsonic.com",
-	"update.nec-display.com",
-	"update.panasonic.com",
-	"update.lg.com",
-	"update.sharp-world.com",
-	"update.hitachi.com",
-	"update.eizo.com",
-	"update.iiyama.com",
-	"update.planar.com",
-	"update.agneovo.com",
-	"update.aoc.com",
-	"update.benq.com",
-	"update.hp.com",
-	"update.philips.com",
-	"update.barco.com",
-	"update.infocus.com",
-	"update.optoma.com",
-	"update.christie.com",
-	"update.nec-display-solutions.com",
-	"update.panasonic.com",
-	"update.sony.com",
-	"update.epson.com",
-	"update.casio.com",
-	"update.mitsubishielectric.com",
-	"update.hitachidigitalmedia.com",
-	"update.canon.com",
-	"update.boxlight.com",
-	"update.vivitekusa.com",
-	"update.viewsonic.com",
-	"update.dell.com",
-	"update.benq.com",
-	"update.infocus.com",
-	"update.optoma.com",
-	"update.hitachi-hk.com.hk",
-	"update.nec-display-solutions.com",
-	"update.panasonic.com",
-	"update.sony.com",
-	"update.casio-intl.com",
-	"update.mitsubishielectric.com",
-	"update.canon.com",
-	"update.christiedigital.com",
-	"update.digitalprojection.com",
-	"update.vivitek.eu",
-	"update.eiki.com",
-	"update.barco.com",
-	"update.sim2.com",
-	"update.projectiondesign.com",
-	"update.projection.panasonic.net",
-	"update.wolfvision.com",
-	"pypi.org",
-	"npmjs.com",
-	"nodejs.org",
-	"golang.org",
-	"github.com",
-	"gitlab.com",
-	"bitbucket.org",
-	"sourceforge.net",
-	"launchpad.net",
-	"pkg.go.dev",
-	"pythonhosted.org",
-	"anaconda.org",
-	"conda-forge.org",
-	"npm.community",
-	"npmjs.org",
-	"yarnpkg.com",
-	"jsdelivr.com",
-	"unpkg.com",
-	"cdnjs.com",
-	"devpi.net",
-	"gopkg.in",
-	"gocenter.io",
-	"proxy.golang.org",
-	"pepy.tech",
-	"libraries.io",
-	"opensourcehub.io",
-	"nodeload.github.com",
-	"raw.githubusercontent.com",
-	"codeload.github.com",
-	"objects.githubusercontent.com",
-	"nodei.co",
-	"badge.fury.io",
-	"gemnasium.com",
-	"npm.runkit.com",
-	"npmjs.cf",
-	"yarnpkg.com",
-	"registry.npmjs.cf",
-	"skimdb.npmjs.com",
-	"npm.pkg.github.com",
-	"packagecloud.io",
-	"bintray.com",
-	"jfrog.io",
-	"registry.npmjs.org",
-	"npmjs.com",
-	"registry.yarnpkg.com",
-	"codeload.github.com",
-	"api.github.com",
-	"raw.github.com",
-	"gist.github.com",
-	"gist.githubusercontent.com",
-	"gist.githubassets.com",
-	"githubassets.com",
-	"objects-origin.githubusercontent.com",
-	"github-releases.githubusercontent.com",
-	"github-cloud.s3.amazonaws.com",
-	"github-com.s3.amazonaws.com",
-	"github-production-release-asset-2e65be.s3.amazonaws.com",
-	"github-production-repository-file-5c1aeb.s3.amazonaws.com",
-	"github-production-user-asset-6210df.s3.amazonaws.com",
-	"github-production-user-asset-79cafe.s3.amazonaws.com",
-	"avatars.githubusercontent.com",
-	"avatars0.githubusercontent.com",
-	"avatars1.githubusercontent.com",
-	"avatars2.githubusercontent.com",
-	"avatars3.githubusercontent.com",
-	"avatars4.githubusercontent.com",
-	"avatars5.githubusercontent.com",
-	"avatars6.githubusercontent.com",
-	"avatars7.githubusercontent.com",
-	"avatars8.githubusercontent.com",
-	"camo.githubusercontent.com",
-	"cloud.githubusercontent.com",
-	"repository-images.githubusercontent.com",
-	"user-images.githubusercontent.com",
-	"gitlab.com",
-	"assets.gitlab-static.net",
-	"gitlab.com/assets",
-	"storage.googleapis.com",
-	"www.gstatic.com",
-	"gitlab-static.net",
-	"bitbucket.org",
-	"api.bitbucket.org",
-	"akamai.bintray.com",
-	"dl.bintray.com",
-	"jcenter.bintray.com",
-	"maven.pkg.github.com",
-	"golang.org",
-	"go.googlesource.com",
-	"storage.googleapis.com/golang",
-	"dl.google.com/go",
-	"storage.googleapis.com",
-	"googleapis.com",
-	"apache.bintray.com",
-	"dl.google.com",
-	"maven.google.com",
-	"pypi.python.org",
-	"files.pythonhosted.org",
-	"pypi.io",
-	"warehouse-cdn.herokuapp.com",
-	"pypi.fury.io",
-	"badge.fury.io",
-	"pypi.org/packages",
-	"files.pythonhosted.org",
-	"pythonhosted.org",
-	"conda.anaconda.org",
-	"repo.anaconda.com",
-	"repo.continuum.io",
-	"conda.anaconda.org",
-	"conda-forge.org",
-	"cdn.jsdelivr.net",
-	"data.jsdelivr.com",
-	"fastly.jsdelivr.net",
-	"gcore.jsdelivr.net",
-	"cdn.jsdelivr.net",
-	"unpkg.com",
-	"registry.npmjs.org",
-	"npm.pkg.github.com",
-	"npm.pkg.github.io",
-	"npm.taobao.org",
-	"npm.yarnpkg.com",
-	"registry.yarnpkg.com",
-	"yarnpkg.com",
-	"yarn.npmjs.org",
-	"unpkg.com",
-	"cdn.jsdelivr.net",
-	"fastly.jsdelivr.net",
-	"package.openupm.com",
-	"npm.pkg.github.com",
-	"registry.npmjs.org",
-	"unpkg.com",
-	"cdn.jsdelivr.net",
-	"npm.taobao.org",
-	"npm.yarnpkg.com",
-	"registry.yarnpkg.com",
-	"yarn.npmjs.org",
-	"raw.githubusercontent.com",
-	"codeload.github.com",
-	"nodeload.github.com",
-	"objects.githubusercontent.com",
-	"pkg.go.dev",
-	"proxy.golang.org",
-	"storage.googleapis.com/go-boringcrypto",
-	"storage.googleapis.com/download.golang.org",
-	"golang.google.cn/dl",
-	"gomirrors.org",
-	"goproxy.cn",
-	"gocenter.io",
-	"mirrors.aliyun.com/goproxy",
-	"goproxy.io",
-	"goproxy.cn",
-	"mirror.go-repo.io",
-	"golang.org/x/build/repos",
-	"proxy.golang.com.cn",
-	"sum.golang.google.cn",
-	"sum.golang.org",
-	"index.golang.org",
-	"godoc.org",
-	"go-search.org",
-	"api.godoc.org",
-	"golang.org/x/pkgsite",
-	"go.googlesource.com/go",
-	"go-review.googlesource.com",
-	"go.googlesource.com/proposal",
-	"go.googlesource.com/gofrontend",
-	"go.googlesource.com/net",
-	"proxy.golang.org",
-	"sum.golang.org",
-	"index.golang.org",
-	"godoc.org",
-	"pkg.go.dev",
-	"go.dev",
-	"go.googlesource.com",
-	"go-review.googlesource.com",
-	"gotools.org",
-	"labix.org/gocheck",
-	"gopkg.in",
-	"gopkg.in/check.v1",
-	"gopkg.in/yaml.v2",
-	"gopkg.in/yaml.v3",
-	"gopkg.in/mgo.v2",
-	"gopkg.in/mgo.v2/bson",
-	"gopkg.in/mgo.v2/txn",
-	"gopkg.in/mgo.v2/internal/sasl",
-	"gopkg.in/mgo.v2/internal/scram",
-	"gopkg.in/mgo.v2/internal/json",
-	"gopkg.in/dgrijalva/jwt-go.v3",
-	"gopkg.in/square/go-jose.v2",
-	"gopkg.in/square/go-jose.v2/json",
-	"gopkg.in/square/go-jose.v2/jwt",
-	"gopkg.in/warnings.v0",
-	"gopkg.in/src-d/go-git.v4",
-	"gopkg.in/src-d/go-git.v4/plumbing",
-	"gopkg.in/src-d/go-git.v4/plumbing/format/gitignore",
-	"gopkg.in/src-d/go-git.v4/plumbing/format/idxfile",
-	"gopkg.in/src-d/go-git.v4/plumbing/format/objfile",
-	"gopkg.in/src-d/go-git.v4/plumbing/format/packfile",
-	"gopkg.in/src-d/go-git.v4/plumbing/format/pktline",
-	"gopkg.in/src-d/go-billy.v4",
-	"gopkg.in/src-d/go-billy.v4/memfs",
-	"gopkg.in/src-d/go-billy.v4/osfs",
-	"hub.docker.com",
-	"quay.io",
-	"gcr.io",
-	"registry.gitlab.com",
-	"ghcr.io",
-	"public.ecr.aws",
-	"mcr.microsoft.com",
-	"registry.redhat.io",
-	"registry.access.redhat.com",
-	"registry.connect.redhat.com",
-	"registry.centos.org",
-	"ftp.debian.org",
-	"download.opensuse.org",
-	"registry.opensuse.org",
-	"registry.fedoraproject.org",
-	"ftp.halifax.rwth-aachen.de",
-	"mirror.gcr.io",
-	"k8s.gcr.io",
-	"registry.k8s.io",
-	"container-registry.oracle.com",
-	"container.cloud.ibm.com",
-	"registry.connect.redhat.com",
-	"registry.redhat.com",
-	"flatcar-linux.org",
-	"tectonic.release.core-os.net",
-	"registry.suse.com",
-	"storage.googleapis.com",
-	"gcr.azk8s.cn",
-	"mirror.azk8s.cn",
-	"archlinux.org",
-	"mirror.rackspace.com",
-	"registry.speedpartner.de",
-	"mirror.aarnet.edu.au",
-	"mirror.steadfast.net",
-	"registry.jetstack.net",
-	"eu.gcr.io",
-	"asia.gcr.io",
-	"us.gcr.io",
-	"nvcr.io",
-	"dockerhub.azk8s.cn",
-	"dockerhub.grid.tf",
-	"cache.ruby-lang.org",
-	"repo.huaweicloud.com",
-	"mirrors.ustc.edu.cn",
-	"mirror.azure.cn",
-	"registry.docker-cn.com",
-	"mirrors.tencent.com",
-	"mirrors.cloud.tencent.com",
-	"hub-mirror.c.163.com",
-	"mirror.ccs.tencentyun.com",
-	"registry.cn-beijing.aliyuncs.com",
-	"registry.cn-hangzhou.aliyuncs.com",
-	"registry.cn-shanghai.aliyuncs.com",
-	"registry.cn-shenzhen.aliyuncs.com",
-	"registry.cn-hongkong.aliyuncs.com",
-	"mirror.lzu.edu.cn",
-	"mirror.bjtu.edu.cn",
-	"mirror.nju.edu.cn",
-	"mirrors.huaweicloud.com",
-	"registry.kaiyuanshe.cn",
-	"mirror.sjtu.edu.cn",
-	"mirrors.nju.edu.cn",
-	"mirror.azure.cn",
-	"mirror.cqu.edu.cn",
-	"mirrors.neusoft.edu.cn",
-	"mirror.lzu.edu.cn",
-	"mirrors.geekpie.club",
-	"mirrors.njupt.edu.cn",
-	"mirror.iscas.ac.cn",
-	"mirror.bit.edu.cn",
-	"mirrors.bfsu.edu.cn",
-	"ftp.sjtu.edu.cn",
-	"mirror.bjtu.edu.cn",
-	"mirrors.hit.edu.cn",
-	"mirrors.dgut.edu.cn",
-	"mirrors.aliyun.com",
-	"mirrors.cqu.edu.cn",
-	"mirrors.yun-idc.com",
-	"mirrors.tuna.tsinghua.edu.cn",
-	"mirror.lzu.edu.cn",
-	"mirrors.neusoft.edu.cn",
-	"mirror.bjtu.edu.cn",
-	"registry.docker-cn.com",
-	"mirror.azure.cn",
-	"mirrors.shu.edu.cn",
-	"docker.mirrors.sjtug.sjtu.edu.cn",
-	"mirror.nju.edu.cn",
-	"mirrors.zju.edu.cn",
-	"mirror.iscas.ac.cn",
-	"mirror.xtom.com.hk",
-	"mirror.xtom.com",
-	"azk8spull.aks.xycloud.com",
-	"mirror.xtom.com.cn",
-	"mirror.kaiyuanshe.cn",
-	"mirrors.xjtu.edu.cn",
-	"mirror.nwafu.edu.cn",
-	"mirror.jlu.edu.cn",
-	"mirror.hnust.cn",
-	"mirrors.bupt.edu.cn",
-	"mirror.bit.edu.cn",
-	"mirrors.sustech.edu.cn",
-	"mirrors.pku.edu.cn",
-	"mirrors.dlut.edu.cn",
-	"mirror.sjtu.edu.cn",
-	"docker.nju.edu.cn",
-	"mirrors.skyshe.cn",
-	"mirrors.njtech.edu.cn",
-	"docker.mirrors.sjtug.sjtu.edu.cn",
-	"mirror.jstvo.com",
-	"mirrors.tongji.edu.cn",
-	"mirror.lzu.edu.cn",
-	"mirrors.ustc.edu.cn",
-	"mirrors.njupt.edu.cn",
-	"mirror.iscas.ac.cn",
-	"mirrors.aliyun.com",
-	"registry.jiangxingai.com",
-	"mirror.bit.edu.cn",
-	"mirrors.hit.edu.cn",
-	"mirrors.cqu.edu.cn",
-	"mirrors.neusoft.edu.cn",
-	"mirror.bjtu.edu.cn",
-	"mirror.azure.cn",
-	"mirror.dlut.edu.cn",
-	"mirror.hnust.cn",
-	"mirror.nju.edu.cn",
-	"mirror.lzu.edu.cn",
-	"mirror.sjtu.edu.cn",
-	"mirror.xtom.com.hk",
-	"mirror.xtom.com.cn",
-	"mirror.xtom.com",
-	"mirror.cqu.edu.cn",
-	"mirror.dgut.edu.cn",
-	"mirror.nwafu.edu.cn",
-	"mirror.bjtu.edu.cn",
-	"mirror.bit.edu.cn",
-	"mirror.shu.edu.cn",
-	"mirror.tuna.tsinghua.edu.cn",
-	"mirror.bfsu.edu.cn",
-	"mirrors.ustc.edu.cn",
-	"mirror.lzu.edu.cn",
-	"azk8spull.aks.xycloud.com",
-	"mirror.jlu.edu.cn",
-	"mirror.njtech.edu.cn",
-	"mirror.jstvo.com",
-	"registry.docker-cn.com",
-	"mirror.iscas.ac.cn",
-	"mirror.hnust.cn",
-	"mirror.cqu.edu.cn",
-	"mirrors.xjtu.edu.cn",
-	"mirrors.neusoft.edu.cn",
-	"mirror.nju.edu.cn",
-	"mirror.bupt.edu.cn",
-	"mirror.shu.edu.cn",
-	"mirror.sjtu.edu.cn",
-	"mirror.hit.edu.cn",
-	"mirror.yun-idc.com",
-	"mirror.njtech.edu.cn",
-	"mirror.xtom.com",
-	"mirror.xtom.com.hk",
-	"mirror.xtom.com.cn",
-	"mirror.jstvo.com",
-	"mirror.bit.edu.cn",
-	"mirrors.nju.edu.cn",
-	"mirror.tongji.edu.cn",
-	"mirrors.sustech.edu.cn",
-	"mirror.jlu.edu.cn",
-	"mirror.njupt.edu.cn",
-	"mirror.pku.edu.cn",
-	"mirror.zju.edu.cn",
-	"mirror.hnust.cn",
-	"mirror.kaiyuanshe.cn",
-	"mirror.skyshe.cn",
-	"mirror.geekpie.club",
-	"docker.mirrors.sjtug.sjtu.edu.cn",
-	"mirror.bjtu.edu.cn",
-	"mirror.nwafu.edu.cn",
-	"mirror.bfsu.edu.cn",
-	"mirror.lzu.edu.cn",
-	"mirror.dlut.edu.cn",
-	"registry.jiangxingai.com",
-	"ftp.sjtu.edu.cn",
-	"mirror.neusoft.edu.cn",
-	"docker.nju.edu.cn",
-	"mirror.bupt.edu.cn",
-	"mirror.azure.cn",
-	"mirrors.aliyun.com",
-	"mirror.dgut.edu.cn",
-	"mirror.shu.edu.cn",
-	"mirror.cqu.edu.cn",
-	"mirror.sjtu.edu.cn",
-	"mirror.hit.edu.cn",
-	"mirror.xtom.com.hk",
-	"mirror.xtom.com.cn",
-	"mirror.xtom.com",
-	"mirror.tuna.tsinghua.edu.cn",
-	"mirror.xjtu.edu.cn",
-	"mirror.njtech.edu.cn",
-	"mirror.jstvo.com",
-	"mirror.ustc.edu.cn",
-	"mirror.lzu.edu.cn",
-	"github.com",
-	"gitlab.com",
-	"bitbucket.org",
-	"sourceforge.net",
-	"launchpad.net",
-	"assembla.com",
-	"azure.microsoft.com",
-	"code.google.com",
-	"codeberg.org",
-	"codeplex.com",
-	"savannah.gnu.org",
-	"repo.or.cz",
-	"git.kernel.org",
-	"git.gnome.org",
-	"git.kde.org",
-	"anongit.freedesktop.org",
-	"git.samba.org",
-	"git.openstack.org",
-	"git.linaro.org",
-	"git.kernel.dk",
-	"git.postgresql.org",
-	"git.zx2c4.com",
-	"git.openwrt.org",
-	"git.lede-project.org",
-	"git.libav.org",
-	"git.ffmpeg.org",
-	"git.videolan.org",
-	"git.busybox.net",
-	"git.buildroot.net",
-	"git.musl-libc.org",
-	"git.uclibc.org",
-	"git.alpinelinux.org",
-	"git.openbsd.org",
-	"git.netbsd.org",
-	"git.freebsd.org",
-	"git.dragonflybsd.org",
-	"cgit.freedesktop.org",
-	"gitweb.freedesktop.org",
-	"git.ti.com",
-	"git.baserock.org",
-	"git.yoctoproject.org",
-	"git.sailfishos.org",
-	"git.linuxfoundation.org",
-	"git.eclipse.org",
-	"git.apache.org",
-	"git.nginx.com",
-	"git.sr.ht",
-	"foss.heptapod.net",
-	"gitlab.freedesktop.org",
-	"gitlab.gnome.org",
-	"source.codeaurora.org",
-	"opendev.org",
-	"github.osuosl.org",
-	"github.riscv.org",
-	"salsa.debian.org",
-	"git.drupalcode.org",
-	"git.fsfe.org",
-	"gitlab.com/gitlab-org",
-	"framagit.org",
-	"git.suckless.org",
-	"git.redox-os.org",
-	"git.haiku-os.org",
-	"gitea.com",
-	"gitgud.io",
-	"invent.kde.org",
-	"notabug.org",
-	"git.launchpad.net",
-	"pagure.io",
-	"repo.progsbase.com",
-	"repo.or.cz",
-	"git.plume-creator.com",
-	"code.wireshark.org",
-	"git.qt.io",
-	"invent.kde.org",
-	"patchwork.ozlabs.org",
-	"android.googlesource.com",
-	"git.blender.org",
-	"gitcafe.com",
-	"gitorious.org",
-	"repo.or.cz",
-	"git.collabora.com",
-	"repo.palemoon.org",
-	"gitlab.com/inkscape",
-	"gitlab.haskell.org",
-	"gitlab.com/mbeddr",
-	"git.netfilter.org",
-	"git.dpdk.org",
-	"git.opendaylight.org",
-	"git.fd.io",
-	"git.qemu.org",
-	"git.xfce.org",
-	"git.lxde.org",
-	"gitlab.com/wireshark",
-	"dev.gentoo.org",
-	"git.centos.org",
-	"git.altlinux.org",
-	"pkgs.fedoraproject.org",
-	"git.guix.gnu.org",
-	"gitlab.archlinux.org",
-	"gitweb.gentoo.org",
-	"git.deepin.io",
-	"git.ghostscript.com",
-	"git.libreoffice.org",
-	"gerrit.libreoffice.org",
-	"git.videolan.org",
-	"git.kodi.tv",
-	"git.mozilla.org",
-	"github.com/gnuradio",
-	"github.com/beagleboard",
-	"github.com/machinekit",
-	"git.openmoko.org",
-	"code.nsnam.org",
-	"git.osmocom.org",
-	"source.denx.de",
-	"git.mbed.com",
-	"git.ti.com",
-	"git.toradex.com",
-	"gitorious.org",
-	"git.linaro.org",
-	"git.zephyrproject.org",
-	"git.denx.de",
-	"git.lwn.net",
-	"github.com/ceph",
-	"git.openstack.org",
-	"git.postgresql.org",
-	"git.infradead.org",
-	"git.kernel.org",
-	"git.savannah.gnu.org",
-	"git.yoctoproject.org",
-	"git.enlightenment.org",
-	"patchwork.kernel.org",
-	"git.sel4.systems",
-	"git.minix3.org",
-	"git.shibboleth.net",
-	"github.com/freebsd",
-	"github.com/NetBSD",
-	"cgit.freebsd.org",
-	"anongit.mindrot.org",
-	"git.exherbo.org",
-	"git.adelielinux.org",
-	"git.adelielinux.org",
-	"git.gnupg.org",
-	"gitlab.com/gnutls",
-	"git.mutt.org",
-	"git.xiph.org",
-	"gitlab.isc.org",
-	"github.com/irssi",
-	"git.claws-mail.org",
-	"gitlab.com/procps-ng",
-	"git.savannah.gnu.org",
-	"git.busybox.net",
-	"git.libssh.org",
-	"git.netfilter.org",
-	"git.openvpn.net",
-	"git.uclibc.org",
-	"git.buildroot.net",
-	"git.sv.gnu.org",
-	"git.ghostscript.com",
-	"git.haproxy.org",
-	"git.xfce.org",
-	"git.lxde.org",
-	"gitlab.gnome.org",
-	"git.enlightenment.org",
-	"git.mate-desktop.org",
-	"git.xfce.org",
-	"git.lxde.org",
-	"git.openbox.org",
-	"git.fluxbox.org",
-	"git.kde.org",
-	"git.trinity-desktop.org",
-	"git.cinnamon-desktop.org",
-	"git.lxqt.org",
-	"git.i3wm.org",
-	"gitlab.freedesktop.org",
-	"github.com/awesomeWM",
-	"git.suckless.org",
-	"git.sway.se",
-	"git.enlightenment.org",
-	"git.gnome.org",
-	"gitlab.gnome.org",
-	"git.xfce.org",
-	"git.lxde.org",
-	"github.com/lxde",
-	"git.linuxmint.com",
-	"git.trisquel.info",
-	"git.parabola.nu",
-	"gitlab.com/pureos",
-	"gitlab.alpinelinux.org",
-	"gitlab.com/postmarketOS",
-	"gitlab.com/ubports",
-	"github.com/manjaro-arm",
-	"gitlab.com/nitrux",
-	"gitlab.com/Whonix",
-	"github.com/OpenMandrivaAssociation",
-	"gitlab.com/redox-os",
-	"github.com/SerenityOS",
-	"gitlab.com/osdev-wiki",
-	"github.com/jbruchon",
-	"git.devuan.org",
-	"git.hyperbola.info",
-	"mirrors.kernel.org",
-	"archive.ubuntu.com",
-	"mirror.centos.org",
-	"ftp.debian.org",
-	"mirrors.fedoraproject.org",
-	"mirror.rackspace.com",
-	"ftp.osuosl.org",
-	"mirror.facebook.net",
-	"mirror.hetzner.de",
-	"mirrors.evowise.com",
-	"mirror.steadfast.net",
-	"mirror.cogentco.com",
-	"mirror.wdc1.us.leaseweb.net",
-	"mirror.enzu.com",
-	"mirror.us.leaseweb.net",
-	"mirror.it.leaseweb.net",
-	"mirror.nl.leaseweb.net",
-	"mirror.liquidtelecom.com",
-	"mirror.eu.oneandone.net",
-	"ftp.nluug.nl",
-	"ftp.sunet.se",
-	"mirror.sov.uk.goscomb.net",
-	"ftp.acc.umu.se",
-	"mirror.umd.edu",
-	"mirror.layeronline.com",
-	"mirror.as43289.net",
-	"mirror.kku.ac.th",
-	"ftp.iinet.net.au",
-	"mirror.rise.ph",
-	"mirrors.tuna.tsinghua.edu.cn",
-	"mirror.bjtu.edu.cn",
-	"mirror.azure.cn",
-	"ftp.sjtu.edu.cn",
-	"ftp.cuhk.edu.hk",
-	"ftp.jaist.ac.jp",
-	"mirrors.ustc.edu.cn",
-	"mirror.nyi.net",
-	"mirror.jre655.com",
-	"mirror.csclub.uwaterloo.ca",
-	"mirror.its.sfu.ca",
-	"mirror.clibre.uqam.ca",
-	"ftp.free.fr",
-	"ftp.oleane.net",
-	"ftp.lip6.fr",
-	"ftp.crihan.fr",
-	"ftp.igh.cnrs.fr",
-	"ftp.ciril.fr",
-	"miroir.univ-lorraine.fr",
-	"ftp.cc.uoc.gr",
-	"ftp.ntua.gr",
-	"ftp.heanet.ie",
-	"ftp.eenet.ee",
-	"ftp.estpak.ee",
-	"ftp.litnet.lt",
-	"mirror.terrahost.no",
-	"ftp.nluug.nl",
-	"ftp.surfnet.nl",
-	"mirror.mirohost.net",
-	"mirror.transip.net",
-	"ftp.rnl.tecnico.ulisboa.pt",
-	"ftp.dei.uc.pt",
-	"ftp.igh.cnrs.fr",
-	"ftp.belnet.be",
-	"ftp.esat.net",
-	"mirror.isoc.org.il",
-	"mirror.poliwangi.ac.id",
-	"kartolo.sby.datautama.net.id",
-	"mirror.aarnet.edu.au",
-	"mirror.lagoon.nc",
-	"ftp.iitm.ac.in",
-	"mirrors.piconets.webwerks.in",
-	"mirror.nkn.in",
-	"mirror.waia.asn.au",
-	"mirror.optus.net",
-	"mirror.amaze.com.au",
-	"mirror.overthewire.com.au",
-	"mirror.intergrid.com.au",
-	"mirror.internode.on.net",
-	"mirror.cse.iitk.ac.in",
-	"mirrors.cse.iitm.ac.in",
-	"mirrors.cse.iitd.ac.in",
-	"mirror.seedhost.com",
-	"mirror.0x.sg",
-	"mirror.aktkn.sg",
-	"mirror.myren.net.my",
-	"mirror.squ.edu.om",
-	"mirror.qnren.qa",
-	"mirror.nren.sy",
-	"mirror.kaust.edu.sa",
-	"mirrors.isu.net.sa",
-	"mirror.marwan.ma",
-	"mirror.tn.ksu.edu.tw",
-	"ftp.cse.yzu.edu.tw",
-	"ftp.twaren.net",
-	"ftp.nctu.edu.tw",
-	"ftp.tku.edu.tw",
-	"ftp.cuhk.edu.hk",
-	"ftp.cs.pu.edu.tw",
-	"ftp.nchc.org.tw",
-	"ftp.cs.ccu.edu.tw",
-	"ftp.mirror.tw",
-	"mirror.pregi.net",
-	"mirror.bytemark.co.uk",
-	"mirror.ox.ac.uk",
-	"mirror.netcologne.de",
-	"mirror.softaculous.com",
-	"mirrors.ocf.berkeley.edu",
-	"mirror.cs.uwp.edu",
-	"mirror.us-midwest-1.nexcess.net",
-	"mirror.hostduplex.com",
-	"ftp.ussg.iu.edu",
-	"mirrors.syringanetworks.net",
-	"repos.lax.quadranet.com",
-	"mirror.pit.teraswitch.com",
-	"mirrors.acm.wpi.edu",
-	"mirrors.bloomu.edu",
-	"ftp.osuosl.org",
-	"mirror.us-midwest-1.nexcess.net",
-	"mirror.sesp.northwestern.edu",
-	"mirror.thelinuxfix.com",
-	"mirrors.rit.edu",
-	"mirror.vtti.vt.edu",
-	"mirror.math.princeton.edu",
-	"pub.allbsd.org",
-	"mirrors.evowise.com",
-	"mirror.steadfast.net",
-	"mirror.cogentco.com",
-	"mirrors.advancedhosters.com",
-	"mirrors.gigenet.com",
-	"mirror.stjschools.org",
-	"mirror.team-cymru.com",
-	"mirror.siena.edu",
-	"mirror.nexcess.net",
-	"mirror.metrocast.net",
-	"ftp.linux.ncsu.edu",
-	"mirror.umd.edu",
-	"mirrors.xmission.com",
-	"mirror.us.leaseweb.net",
-	"mirror.jmu.edu",
-	"mirror.cc.columbia.edu",
-	"mirror.es.its.nyu.edu",
-	"ftp.gtlib.gatech.edu",
-	"ftp.sunet.se",
-	"mirror.sov.uk.goscomb.net",
-	"ftp.yzu.edu.tw",
-	"ftp.jaist.ac.jp",
-	"mirrors.neusoft.edu.cn",
-	"mirror.bjtu.edu.cn",
-	"mirror.bfsu.edu.cn",
-	"mirrors.163.com",
-	"mirror.opencas.org",
-	"mirror.devoli.net",
-	"mirrors.dotsrc.org",
-	"ftp.halifax.rwth-aachen.de",
-	"mirror.23media.de",
-	"mirror.eu.oneandone.net",
-	"ftp.rrzn.uni-hannover.de",
-	"ftp.fau.de",
-	"ftp-stud.hs-esslingen.de",
-	"mirror.de.leaseweb.net",
-	"debian.charite.de",
-	"debian.tu-bs.de",
-	"ftp.uni-kl.de",
-	"ftp.uni-bayreuth.de",
-	"ftp.tu-clausthal.de",
-	"ftp.informatik.hu-berlin.de",
-	"mirror.easyname.at",
-	"mirror.kumi.systems",
-	"mirror.pregi.net",
-	"ftp5.gwdg.de",
-	"mirror.dogado.de",
-	"ftp.rz.tu-bs.de",
-	"ftp.tu-chemnitz.de",
-	"ftp.rrze.uni-erlangen.de",
-	"mirror.23media.com",
-	"mirror.gnomus.de",
-	"mirror.serverloft.eu",
-	"packages.hs-regensburg.de",
-	"mirror-fpt-telecom.fpt.net",
-	"mirror.viettel.vn",
-	"mirrors.vinahost.vn",
-	"mirrors.digipower.vn",
-	"mirror.morgancyber.net",
-	"mirror.smartnetworks.net.in",
-	"mirror.amsiohosting.net",
-	"ftp.icm.edu.pl",
-	"ftp.man.poznan.pl",
-	"ftp.piotrkosoft.net",
-	"ftp.ps.pl",
-	"mirror.onet.pl",
-	"ftp.vectranet.pl",
-	"mirror.serverion.com",
-	"ftp.task.gda.pl",
-	"ftp.agh.edu.pl",
-	"mirror.chmuri.net",
-	"mirror.kumi.systems",
-	"mirrors.dotsrc.org",
-	"ftp.uni-bayreuth.de",
-	"mirror.united-gameserver.de",
-	"mirror.de.leaseweb.net",
-	"security.ubuntu.com",
-	"archive.ubuntu.com",
-	"mirrors.ubuntu.com",
-	"us.archive.ubuntu.com",
-	"apt.releases.hashicorp.com",
-	"mirrors.fedoraproject.org",
-	"download.fedoraproject.org",
-	"ftp.redhat.com",
-	"mirrors.kernel.org",
-	"yum.oracle.com",
-	"mirror.centos.org",
-	"mirrors.centos.org",
-	"vault.centos.org",
-	"download.opensuse.org",
-	"mirror.opensuse.org",
-	"updates.opensuse.org",
-	"packages.microsoft.com",
-	"dl.fedoraproject.org",
-	"dl.google.com",
-	"mirrors.ustc.edu.cn",
-	"mirror.bytemark.co.uk",
-	"ftp.fau.de",
-	"mirrors.edge.kernel.org",
-	"ftp.jaist.ac.jp",
-	"mirror.yandex.ru",
-	"mirrors.aliyun.com",
-	"mirrors.tuna.tsinghua.edu.cn",
-	"mirrors.huaweicloud.com",
-	"mirrors.163.com",
-	"repo.huaweicloud.com",
-	"mirror.rackspace.com",
-	"ftp.cuhk.edu.hk",
-	"ftp.sjtu.edu.cn",
-	"mirrors.zju.edu.cn",
-	"ftp.riken.jp",
-	"ftp.yz.yamagata-u.ac.jp",
-	"mirror.nforce.com",
-	"mirror.aarnet.edu.au",
-	"ftp.iitm.ac.in",
-	"www.mirrorservice.org",
-	"mirror.hetzner.de",
-	"mirrors.dotsrc.org",
-	"mirror.lzu.edu.cn",
-	"mirrors.bupt.edu.cn",
-	"mirror.bjtu.edu.cn",
-	"mirror.neu.edu.cn",
-	"mirror.bit.edu.cn",
-	"mirrors.bfsu.edu.cn",
-	"mirrors.cqu.edu.cn",
-	"mirror.iscas.ac.cn",
-	"mirror.lzu6.edu.cn",
-	"mirror.nju.edu.cn",
-	"mirrors.hit.edu.cn",
-	"mirrors.dgut.edu.cn",
-	"mirror.sjtu.edu.cn",
-	"mirror.xtom.com.hk",
-	"mirror.se.sjtu.edu.cn",
-	"ftp.sjtu.edu.cn",
-	"mirror.azure.cn",
-	"mirror.pcbeta.com",
-	"mirror.newasia.cloud",
-	"mirrors.njupt.edu.cn",
-	"mirror.sdu.edu.cn",
-	"mirror.dlut.edu.cn",
-	"mirror.jlu.edu.cn",
-	"mirrors.tsinghua.edu.cn",
-	"mirrors.cqu.edu.cn",
-	"mirror.bit.edu.cn",
-	"mirrors.njupt.edu.cn",
-	"mirror.neu.edu.cn",
-	"mirrors.ustc.edu.cn",
-	"mirrors.xjtu.edu.cn",
-	"mirror.bjtu.edu.cn",
-	"mirror.lzu.edu.cn",
-	"mirror.nju.edu.cn",
-	"mirror.qdu.edu.cn",
-	"mirrors.zju.edu.cn",
-	"mirrors.neusoft.edu.cn",
-	"mirrors.hust.edu.cn",
-	"mirrors.sjtug.sjtu.edu.cn",
-	"mirror.sustech.edu.cn",
-	"mirrors.tongji.edu.cn",
-	"mirrors.yun-idc.com",
-	"mirror.lzu6.edu.cn",
-	"mirrors.aliyun.com",
-	"mirrors.huaweicloud.com",
-	"mirrors.163.com",
-	"mirror.sjtu.edu.cn",
-	"mirrors.dgut.edu.cn",
-	"mirror.xtom.com.hk",
-	"mirror.dlut.edu.cn",
-	"mirrors.bupt.edu.cn",
-	"mirror.iscas.ac.cn",
-	"mirrors.hit.edu.cn",
-	"mirror.jlu.edu.cn",
-	"mirror.sdu.edu.cn",
-	"mirror.nku.edu.tw",
-	"ftp.twaren.net",
-	"mirror.yzu.edu.tw",
-	"mirror.nchc.org.tw",
-	"mirror.cse.iitk.ac.in",
-	"mirror.metrocast.net",
-	"mirror.pregi.net",
-	"ftp.daum.net",
-	"mirror.kakao.com",
-	"mirror.navercorp.com",
-	"ftp.kaist.ac.kr",
-	"ftp.tsukuba.wide.ad.jp",
-	"ftp.nara.wide.ad.jp",
-	"ftp.kddilabs.jp",
-	"ftp.ring.gr.jp",
-	"mirror.fairway.ne.jp",
-	"mirror.onet.pl",
-	"ftp.icm.edu.pl",
-	"ftp.pwr.wroc.pl",
-	"ftp.vectranet.pl",
-	"mirror.yandex.ru",
-	"mirror.truenetwork.ru",
-	"mirror.corbina.net",
-	"mirror.docker.ru",
-	"mirrors.powernet.com.ru",
-	"mirror.timeweb.ru",
-	"mirror.mephi.ru",
-	"mirror.logol.ru",
-	"mirror.kamtv.ru",
-	"mirror.awanti.com",
-	"mirror.bezdrat.net.ua",
-	"mirror.mirohost.net",
-	"mirror.zyrianovsk.net",
-	"ftp.byfly.by",
-	"mirror.datacenter.by",
-	"ftp.kuzstu.ru",
-	"mirror.kuzstu.ru",
-	"mirror.lnk.lv",
-	"mirror.linux.ee",
-	"ftp.aso.ee",
-	"ftp.heanet.ie",
-	"ftp.eenet.ee",
-	"mirror.as43289.net",
-	"mirrors.netix.net",
-	"mirror.koddos.net",
-	"mirror.vpsnet.com",
-	"mirror.init7.net",
-	"pkg.adfinis.com",
-	"mirror.switch.ch",
-	"pkg.bfh.ch",
-	"mirror.solnet.ch",
-	"mirror.solnet.ch",
-	"ubuntu.ethz.ch",
-	"ftp.belnet.be",
-	"mirror.as35701.net",
-	"mirror.transip.net",
-	"ftp.nluug.nl",
-	"mirror.nl.leaseweb.net",
-	"mirror.previder.nl",
-	"mirror.eu.oneandone.net",
-	"nl.archive.ubuntu.com",
-	"mirror.de.leaseweb.net",
-	"ftp.halifax.rwth-aachen.de",
-	"mirror.23media.de",
-	"mirror.netcologne.de",
-	"mirror.satis-faction.de",
-	"mirror.serverloft.eu",
-	"mirror.dogado.de",
-	"mirror.metalgamer.eu",
-	"mirror.de.gaw.wtnet.de",
-	"mirror.wtnet.de",
-	"mirror.23media.com",
-	"ftp.uni-erlangen.de",
-	"ftp.gwdg.de",
-	"ftp.fau.de",
-	"ftp-stud.hs-esslingen.de",
-	"mirror.united-gameserver.de",
-	"ftp.plusline.net",
-	"download.hs-augsburg.de",
-	"ftp.tu-chemnitz.de",
-	"ftp.uni-bayreuth.de",
-	"mirror.informatik.hs-fulda.de",
-	"ftp.rrzn.uni-hannover.de",
-	"ftp.informatik.uni-frankfurt.de",
-	"ftp.uni-kl.de",
-	"mirror.netcologne.de",
-	"mirrors.n-ix.net",
-	"debian.charite.de",
-	"ftp.freenet.de",
-	"ftp.fu-berlin.de",
-	"ftp.uni-stuttgart.de",
-	"ftp.tu-clausthal.de",
-	"ftp.uni-mainz.de",
-	"mirror.softaculous.com",
-	"mirror.appliedi.net",
-	"mirrors.xtom.com",
-	"mirror.team-cymru.com",
-	"mirrors.up.pt",
-	"mirror.pop-sc.rnp.br",
-	"mirror.unesp.br",
-	"mirrors.ufam.edu.br",
-	"mirror.venturasystems.tech",
-	"mirror.ufro.cl",
-	"mirror.uchile.cl",
-	"update.microsoft.com.nsatc.net",
-	"windowsupdate.com.c.163.com",
-	"au.download.windowsupdate.com.c.163.com",
-	"tlu.dl.delivery.mp.microsoft.com.c.163.com",
-	"mscrl.microsoft.com.c.163.com",
-	"dl.delivery.mp.microsoft.com.c.163.com",
-	"update.microsoft.com.gtmc.nl.uu.net",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"update.microsoft.com.nsatc.net",
-	"update.microsoft.com.nsatc.net",
-	"win10.ipv6.microsoft.com.c.163.com",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"windowsupdate.com.nsatc.net",
-	"windowsupdate.microsoft.com.c.163.com",
-	"au.v4.download.windowsupdate.com.c.163.com",
-	"geover-prod.do.dsp.mp.microsoft.com.c.163.com",
-	"update.microsoft.com.nsatc.net",
-	"win10.ipv6.microsoft.com.c.163.com",
-	"origin-www.microsoft.com.c.163.com",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"update.microsoft.com.nsatc.net",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"windowsupdate.com.c.163.com",
-	"mscrl.microsoft.com.c.163.com",
-	"update.microsoft.com.gtmc.nl.uu.net",
-	"ctldl.windowsupdate.com.c.163.com",
-	"windowsupdate.com.c.163.com",
-	"au.download.windowsupdate.com.c.163.com",
-	"win10.ipv6.microsoft.com.nsatc.net",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"au.download.windowsupdate.com.c.163.com",
-	"mscrl.microsoft.com.c.163.com",
-	"dl.delivery.mp.microsoft.com.c.163.com",
-	"au.download.windowsupdate.com.nsatc.net",
-	"windowsupdate.com.nsatc.net",
-	"update.microsoft.com.nsatc.net",
-	"au.download.windowsupdate.com.nsatc.net",
-	"win10.ipv6.microsoft.com.nsatc.net",
-	"tlu.dl.delivery.mp.microsoft.com.c.163.com",
-	"dl.delivery.mp.microsoft.com.c.163.com",
-	"win10.ipv6.microsoft.com.c.163.com",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"ctldl.windowsupdate.com.nsatc.net",
-	"ctldl.windowsupdate.com.c.163.com",
-	"win10.ipv6.microsoft.com.nsatc.net",
-	"windowsupdate.com.c.163.com",
-	"update.microsoft.com.nsatc.net",
-	"dl.delivery.mp.microsoft.com.nsatc.net",
-	"dl.delivery.mp.microsoft.com.c.163.com",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"update.microsoft.com.nsatc.net",
-	"win10.ipv6.microsoft.com.c.163.com",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"windowsupdate.com.nsatc.net",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"win10.ipv6.microsoft.com.nsatc.net",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"au.download.windowsupdate.com.nsatc.net",
-	"dl.delivery.mp.microsoft.com.c.163.com",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"au.v4.download.windowsupdate.com.c.163.com",
-	"ctldl.windowsupdate.com.c.163.com",
-	"update.microsoft.com.nsatc.net",
-	"au.download.windowsupdate.com.nsatc.net",
-	"mscrl.microsoft.com.c.163.com",
-	"au.download.windowsupdate.com.c.163.com",
-	"dl.delivery.mp.microsoft.com.nsatc.net",
-	"win10.ipv6.microsoft.com.c.163.com",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"update.microsoft.com.nsatc.net",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"windowsupdate.com.c.163.com",
-	"update.microsoft.com.nsatc.net",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"dl.delivery.mp.microsoft.com.c.163.com",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"win10.ipv6.microsoft.com.nsatc.net",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"windowsupdate.com.c.163.com",
-	"ctldl.windowsupdate.com.c.163.com",
-	"win10.ipv6.microsoft.com.c.163.com",
-	"update.microsoft.com.nsatc.net",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"dl.delivery.mp.microsoft.com.c.163.com",
-	"update.microsoft.com.nsatc.net",
-	"au.download.windowsupdate.com.c.163.com",
-	"au.download.windowsupdate.com.nsatc.net",
-	"ctldl.windowsupdate.com.nsatc.net",
-	"win10.ipv6.microsoft.com.c.163.com",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"dl.delivery.mp.microsoft.com.c.163.com",
-	"windowsupdate.com.c.163.com",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"au.v4.download.windowsupdate.com.c.163.com",
-	"update.microsoft.com.nsatc.net",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"au.download.windowsupdate.com.c.163.com",
-	"windowsupdate.com.nsatc.net",
-	"dl.delivery.mp.microsoft.com.nsatc.net",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"win10.ipv6.microsoft.com.nsatc.net",
-	"update.microsoft.com.nsatc.net",
-	"mscrl.microsoft.com.c.163.com",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"dl.delivery.mp.microsoft.com.c.163.com",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"win10.ipv6.microsoft.com.c.163.com",
-	"update.microsoft.com.nsatc.net",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"au.download.windowsupdate.com.nsatc.net",
-	"dl.delivery.mp.microsoft.com.c.163.com",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"dl.delivery.mp.microsoft.com.nsatc.net",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"update.microsoft.com.nsatc.net",
-	"win10.ipv6.microsoft.com.nsatc.net",
-	"mscrl.microsoft.com.c.163.com",
-	"dl.delivery.mp.microsoft.com.c.163.com",
-	"win10.ipv6.microsoft.com.c.163.com",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"tlu.dl.delivery.mp.microsoft.com.c.163.com",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"win10.ipv6.microsoft.com.c.163.com",
-	"update.microsoft.com.nsatc.net",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"dl.delivery.mp.microsoft.com.nsatc.net",
-	"update.microsoft.com.nsatc.net",
-	"ctldl.windowsupdate.com.nsatc.net",
-	"win10.ipv6.microsoft.com.nsatc.net",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"dl.delivery.mp.microsoft.com.c.163.com",
-	"au.download.windowsupdate.com.c.163.com",
-	"au.download.windowsupdate.com.nsatc.net",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"update.microsoft.com.nsatc.net",
-	"dl.delivery.mp.microsoft.com.c.163.com",
-	"mscrl.microsoft.com.c.163.com",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"windowsupdate.com.c.163.com",
-	"dl.delivery.mp.microsoft.com.nsatc.net",
-	"update.microsoft.com.nsatc.net",
-	"win10.ipv6.microsoft.com.c.163.com",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"windowsupdate.com.nsatc.net",
-	"mscrl.microsoft.com.c.163.com",
-	"win10.ipv6.microsoft.com.nsatc.net",
-	"dl.delivery.mp.microsoft.com.c.163.com",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"update.microsoft.com.nsatc.net",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"update.microsoft.com.nsatc.net",
-	"win10.ipv6.microsoft.com.c.163.com",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"dl.delivery.mp.microsoft.com.c.163.com",
-	"ctldl.windowsupdate.com.c.163.com",
-	"update.microsoft.com.nsatc.net",
-	"win10.ipv6.microsoft.com.nsatc.net",
-	"windowsupdate.com.nsatc.net",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"dl.delivery.mp.microsoft.com.nsatc.net",
-	"update.microsoft.com.nsatc.net",
-	"dl.delivery.mp.microsoft.com.c.163.com",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"update.microsoft.com.nsatc.net",
-	"win10.ipv6.microsoft.com.c.163.com",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"au.download.windowsupdate.com.c.163.com",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"win10.ipv6.microsoft.com.c.163.com",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"dl.delivery.mp.microsoft.com.c.163.com",
-	"update.microsoft.com.nsatc.net",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"win10.ipv6.microsoft.com.nsatc.net",
-	"dl.delivery.mp.microsoft.com.nsatc.net",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"mscrl.microsoft.com.c.163.com",
-	"update.microsoft.com.nsatc.net",
-	"win10.ipv6.microsoft.com.c.163.com",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"update.microsoft.com.nsatc.net",
-	"dl.delivery.mp.microsoft.com.c.163.com",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"au.download.windowsupdate.com.c.163.com",
-	"tlu.dl.delivery.mp.microsoft.com.c.163.com",
-	"update.microsoft.com.nsatc.net",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"win10.ipv6.microsoft.com.c.163.com",
-	"windowsupdate.com.c.163.com",
-	"update.microsoft.com.nsatc.net",
-	"ctldl.windowsupdate.com.nsatc.net",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"au.download.windowsupdate.com.nsatc.net",
-	"dl.delivery.mp.microsoft.com.c.163.com",
-	"win10.ipv6.microsoft.com.nsatc.net",
-	"update.microsoft.com.c.edgecastcdn.net",
-	"appldnld.apple.com",
-	"mesu.apple.com",
-	"gdmf.apple.com",
-	"www.apple.com",
-	"support.apple.com",
-	"developer.apple.com",
-	"itunes.apple.com",
-	"itunesconnect.apple.com",
-	"appleid.apple.com",
-	"ns.itunes.apple.com",
-	"init.itunes.apple.com",
-	"init-p01md.apple.com",
-	"init-p01st.push.apple.com",
-	"xp.apple.com",
-	"bag.itunes.apple.com",
-	"albert.apple.com",
-	"captive.apple.com",
-	"gs.apple.com",
-	"humb.apple.com",
-	"static.ips.apple.com",
-	"seq.apple.com",
-	"su.itunes.apple.com",
-	"updates-http.cdn-apple.com",
-	"updates.cdn-apple.com",
-	"oscdn.apple.com",
-	"osrecovery.apple.com",
-	"osxapps.itunes.apple.com",
-	"ipsw.me",
-	"hanalei1.cdn.apple.com.akadns.net",
-	"itunes.apple.com.edgekey.net",
-	"appldnld.g.aaplimg.com",
-	"appldnld.apple.com.akadns.net",
-	"mesu-cdn.apple.com.akadns.net",
-	"mesu-cdn.origin-apple.com.akadns.net",
-	"swcdn.apple.com",
-	"swdownload.apple.com",
-	"swscan.apple.com",
-	"swdist.apple.com",
-	"swquery.apple.com",
-	"configuration.apple.com",
-	"guzzoni.apple.com",
-	"setup.icloud.com",
-	"p01-keyvalueservice.icloud.com",
-	"p01-bookmarks.icloud.com",
-	"p01-contacts.icloud.com",
-	"p01-calendars.icloud.com",
-	"p01-reminders.icloud.com",
-	"p01-sharedstreams.icloud.com",
-	"p01-streams.icloud.com",
-	"p01-ubiquity.icloud.com",
-	"p01-mailws.icloud.com",
-	"p01-ckdatabase.icloud.com",
-	"p01-docws.icloud.com",
-	"p01-uploadws.icloud.com",
-	"p01-quota.icloud.com",
-	"p01-fmf.icloud.com",
-	"p01-fmip.icloud.com",
-	"p01-keyvalueservice.fe.apple-dns.net",
-	"p01-bookmarks.fe.apple-dns.net",
-	"p01-contacts.fe.apple-dns.net",
-	"p01-calendars.fe.apple-dns.net",
-	"p01-reminders.fe.apple-dns.net",
-	"p01-sharedstreams.fe.apple-dns.net",
-	"p01-streams.fe.apple-dns.net",
-	"p01-ubiquity.fe.apple-dns.net",
-	"p01-mailws.fe.apple-dns.net",
-	"p01-ckdatabase.fe.apple-dns.net",
-	"p01-docws.fe.apple-dns.net",
-	"p01-uploadws.fe.apple-dns.net",
-	"p01-quota.fe.apple-dns.net",
-	"p01-fmf.fe.apple-dns.net",
-	"p01-fmip.fe.apple-dns.net",
-	"lcdn-locator.apple.com",
-	"lcdn-locator.apple.com.akadns.net",
-	"lcdn-registration.apple.com",
-	"pancake.apple.com",
-	"pancake.cdn-apple.com.akadns.net",
-	"valid.apple.com",
-	"valid.origin-apple.com.akadns.net",
-	"pancake.g.aaplimg.com",
-	"gsp10-ssl.ls.apple.com",
-	"gspe35-ssl.ls.apple.com",
-	"gspe1-ssl.ls.apple.com",
-	"gspe85-ssl.ls.apple.com",
-	"gspe19-ssl.ls.apple.com",
-	"gsp4-ssl.ls.apple.com",
-	"gspe11-ssl.ls.apple.com",
-	"gsp3-ssl.ls.apple.com",
-	"gsp64-ssl.ls.apple.com",
-	"gsp85-ssl.ls.apple.com",
-	"gspe12-ssl.ls.apple.com",
-	"gspe19-cn-ssl.ls.apple.com",
-	"gspe19-2-cn-ssl.ls.apple.com",
-	"gspe21-ssl.ls.apple.com",
-	"gspe21-cn-ssl.ls.apple.com",
-	"gspe21-2-cn-ssl.ls.apple.com",
-	"client-api.itunes.apple.com",
-	"inn-itunes.apple.com",
-	"pd.itunes.apple.com",
-	"upp.itunes.apple.com",
-	"play.itunes.apple.com",
-	"se.itunes.apple.com",
-	"se2.itunes.apple.com",
-	"sp.itunes.apple.com",
-	"hooks.apple.com",
-	"buy.itunes.apple.com",
-	"init.gc.apple.com",
-	"syncevent.itunes.apple.com",
-	"iosapps.itunes.apple.com",
-	"init.itunes.apple.com.akadns.net",
-	"init-p01md.apple.com.akadns.net",
-	"init-p01st.push.apple.com.akadns.net",
-	"metrics.apple.com",
-	"dzc.apple.com",
-	"ocsp.apple.com",
-	"certs.apple.com",
-	"certs2.apple.com",
-	"certs.g.aaplimg.com",
-	"crl.apple.com",
-	"crl4.apple.com",
-	"crl3.apple.com",
-	"crl5.apple.com",
-	"crl2.apple.com",
-	"crl.g.aaplimg.com",
-	"crl4.g.aaplimg.com",
-	"crl3.g.aaplimg.com",
-	"crl5.g.aaplimg.com",
-	"crl2.g.aaplimg.com",
-	"devcenter.apple.com",
-	"devforums.apple.com",
-	"discussions.apple.com",
-	"idmsa.apple.com",
-	"identity.apple.com",
-	"iforgot.apple.com",
-	"appleid.cdn-apple.com",
-	"gsa.apple.com",
-	"ppq.apple.com",
-	"daap.itunes.apple.com",
-	"cdn-cn1.apple-mapkit.com",
-	"cdn-cn2.apple-mapkit.com",
-	"cdn-cn3.apple-mapkit.com",
-	"cdn-cn4.apple-mapkit.com",
-	"cdn1.apple-mapkit.com",
-	"cdn2.apple-mapkit.com",
-	"cdn3.apple-mapkit.com",
-	"cdn4.apple-mapkit.com",
-	"api-glb-ams.smoot.apple.com",
-	"api-glb-bos.smoot.apple.com",
-	"api-glb-chi.smoot.apple.com",
-	"api-glb-dfw.smoot.apple.com",
-	"api-glb-lax.smoot.apple.com",
-	"api-glb-iad.smoot.apple.com",
-	"api-glb-nyc.smoot.apple.com",
-	"api-glb-nrt.smoot.apple.com",
-	"api-glb-sao.smoot.apple.com",
-	"api-glb-sea.smoot.apple.com",
-	"api-glb-syd.smoot.apple.com",
-	"api-glb-sin.smoot.apple.com",
-	"api-glb-sjc.smoot.apple.com",
-	"api-glb-waw.smoot.apple.com",
-	"api-glb-yyz.smoot.apple.com",
-	"api-glb-cdg.smoot.apple.com",
-	"api-glb-hkg.smoot.apple.com",
-	"api-glb-lhr.smoot.apple.com",
-	"scv.apple.com",
-	"query.ess.apple.com",
-	"query-geo.ess.apple.com",
-	"query-cn.ess.apple.com",
-	"query-tw.ess.apple.com",
-	"query-hk.ess.apple.com",
-	"query-jp.ess.apple.com",
-	"query-ams.ess.apple.com",
-	"query-ind.ess.apple.com",
-	"query-aus.ess.apple.com",
-	"query-kor.ess.apple.com",
-	"xp.apple.com.edgekey.net",
-	"tb.apple.com",
-	"tbsc.apple.com",
-	"cstat.apple.com",
-	"iadsdk.apple.com",
-	"banner.itunes.apple.com",
-	"phobos.apple.com.edgekey.net",
-	"amp-api.apps.apple.com",
-	"amp-api-search-edge.apps.apple.com",
-	"amp-api-edge.apps.apple.com",
-	"itsliveradio.apple.com",
-	"mvod.itunes.apple.com",
-	"radio-services.itunes.apple.com",
-	"radio-activity.itunes.apple.com",
-	"radio-services-qa.itunes.apple.com",
-	"streamingaudio.itunes.apple.com",
-	"download.developer.apple.com",
-	"app.itunes.apple.com",
-	"beta.itunes.apple.com",
-	"betadownload.apple.com",
-	"iosapps.itunes-apple.com.akadns.net",
-	"osxapps-ssl.itunes.apple.com",
-	"osxapps.itunes.g.aaplimg.com",
-	"osxapps.ssl-itunes.g.aaplimg.com",
-	"images.apple.com",
-	"android.googleapis.com",
-	"connectivitycheck.gstatic.com",
-	"clients3.google.com",
-	"android.clients.google.com",
-	"play.googleapis.com",
-	"mtalk.google.com",
-	"mobilemaps.clients.google.com",
-	"photos.google.com",
-	"googlehosted.l.googleusercontent.com",
-	"lh1.googleusercontent.com",
-	"lh2.googleusercontent.com",
-	"lh3.googleusercontent.com",
-	"lh4.googleusercontent.com",
-	"lh5.googleusercontent.com",
-	"lh6.googleusercontent.com",
-	"lh7.googleusercontent.com",
-	"lh8.googleusercontent.com",
-	"lh9.googleusercontent.com",
-	"googleweblight.com",
-	"gvt1.com",
-	"appspot.com",
-	"gvt2.com",
-	"gvt3.com",
-	"ggpht.com",
-	"googleusercontent.com",
-	"gstatic.com",
-	"google.com",
-	"youtubei.googleapis.com",
-	"youtube.googleapis.com",
-	"oauthaccountmanager.googleapis.com",
-	"content.googleapis.com",
-	"clientservices.googleapis.com",
-	"pack.google.com",
-	"books.google.com",
-	"login.corp.google.com",
-	"payments.google.com",
-	"mail-attachment.googleusercontent.com",
-	"translate.googleapis.com",
-	"googlecode.com",
-	"googleapis.l.google.com",
-	"googleadservices.com",
-	"id.google.com",
-	"SafetyNet.googleapis.com",
-	"maps.googleapis.com",
-	"gcm-http.googleapis.com",
-	"www.googleapis.com",
-	"corp.google.com",
-	"stun.l.google.com",
-	"accounts.youtube.com",
-	"domains.google.com",
-	"dl.google.com",
-	"dl-ssl.google.com",
-	"people-pa.googleapis.com",
-	"goto.google.com",
-	"suggestqueries.google.com",
-	"firebaselogging-pa.googleapis.com",
-	"firebaselogging.googleapis.com",
-	"firebaseinstallations.googleapis.com",
-	"firebasedynamiclinks.googleapis.com",
-	"firebaseremoteconfig.googleapis.com",
-	"peopleapi.googleapis.com",
-	"firebaseappcheck.googleapis.com",
-	"instantmessaging-pa.googleapis.com",
-	"socialhelp.clients6.google.com",
-	"appsbackup-pa.clients6.google.com",
-	"oauthaccountmanager.clients6.google.com",
-	"restore-pa.clients6.google.com",
-	"keep-pa.clients6.google.com",
-	"dmagent.google.com",
-	"android.clients.google.com.ar",
-	"android.clients.google.com.au",
-	"android.clients.google.com.br",
-	"android.clients.google.com.cn",
-	"android.clients.google.com.co",
-	"android.clients.google.com.eg",
-	"android.clients.google.com.hk",
-	"android.clients.google.com.mx",
-	"android.clients.google.com.my",
-	"android.clients.google.com.ng",
-	"android.clients.google.com.pe",
-	"android.clients.google.com.ph",
-	"android.clients.google.com.pk",
-	"android.clients.google.com.sa",
-	"android.clients.google.com.sg",
-	"android.clients.google.com.tr",
-	"android.clients.google.com.tw",
-	"android.clients.google.com.ua",
-	"android.clients.google.com.vn",
-	"www-opensocial.googleusercontent.com",
-	"plusone.google.com",
-	"plus.google.com",
-	"yt3.ggpht.com",
-	"play.google.com",
-	"developer.android.com",
-	"market.android.com",
-	"android-developers.blogspot.com",
-	"blog.google",
-	"support.google.com",
-	"maps.gstatic.com",
-	"maps.google.com",
-	"lh0.ggpht.com",
-	"lh1.ggpht.com",
-	"lh2.ggpht.com",
-	"lh3.ggpht.com",
-	"lh4.ggpht.com",
-	"lh5.ggpht.com",
-	"lh6.ggpht.com",
-	"lh0.googleusercontent.com",
-	"lh10.ggpht.com",
-	"lh10.googleusercontent.com",
-	"storage.googleapis.com",
-	"googleapis.com",
-	"download.tensorflow.org",
-	"services.googleapis.com",
-	"chromiumcodereview.appspot.com",
-	"apis.google.com",
-	"toolbarqueries.clients.google.com",
-	"androidquery.googleapis.com",
-	"codereview.chromium.org",
-	"chromium.googlesource.com",
-	"console.cloud.google.com",
-	"console.firebase.google.com",
-	"console.developers.google.com",
-	"console.actions.google.com",
-	"lens.google.com",
-	"optimizationguide-pa.googleapis.com",
-	"assistant.google.com",
-	"vr.google.com",
-	"tpc.googlesyndication.com",
-	"pagead2.googlesyndication.com",
-	"www.googleadservices.com",
-	"ad.doubleclick.net",
-	"static.doubleclick.net",
-	"adservice.google.com",
-	"ade.googlesyndication.com",
-	"adwords.google.com",
-	"analytics.google.com",
-	"ads.google.com",
-	"ssl-google-analytics.l.google.com",
-	"google-analytics.com",
-	"googleadapis.l.google.com",
-	"remoting-pa.googleapis.com",
-	"gstaticadssl.l.google.com",
-	"partnerad.l.doubleclick.net",
-	"www.googletagservices.com",
-	"www.googletagmanager.com",
-	"www-googletagmanager.l.google.com",
-	"gtm.oasisfeng.com",
-	"firebase-settings.crashlytics.com",
-	"settings.crashlytics.com",
-	"app-measurement.com",
-	"app-measurement-api.com",
-	"googleplaystore.net",
-	"redirector.gvt1.com",
-	"redirector.c.bigcache.googleapis.com",
-	"redirector.csi.gstatic.com",
-	"safebrowsing.google.com",
-	"alt1-safebrowsing.google.com",
-	"alt2-safebrowsing.google.com",
-	"alt3-safebrowsing.google.com",
-	"alt4-safebrowsing.google.com",
-	"alt5-safebrowsing.google.com",
-	"alt6-safebrowsing.google.com",
-	"alt7-safebrowsing.google.com",
-	"alt8-safebrowsing.google.com",
-	"alt9-safebrowsing.google.com",
-	"safebrowsing.googleapis.com",
-	"clientmetrics-pa.googleapis.com",
-	"nearby.google.com",
-	"firebasestorage.googleapis.com",
-	"update.googleapis.com",
-	"smartlock.google.com",
-	"gcs-pa-googleapis.l.google.com",
-	"googlehosts.org",
-	"gstatic.lc",
-	"www.google-analytics.com",
-	"wallet.google.com",
-	"adwords-community.com",
-	"geoar.google.com",
-	"blobcomments.googlegroups.com",
-	"chrome.google.com",
-	"android.com",
-	"appengine.google.com",
-	"archive.google.com",
-	"dev.google.com",
-	"code.google.com",
-	"id.l.google.com",
-	"chromeexperiments.com",
-	"data.google.com",
-	"developer.chrome.com",
-	"earth.google.com",
-	"edutrainingcenter.withgoogle.com",
-	"client-channel.google.com",
-	"chart.apis.google.com",
-	"accounts.google.com",
-	"ajax.googleapis.com",
-	"admin.google.com",
-	"material.io",
-	"design.google",
-	"privacy.google.com",
-}
-
-var mailHosts = []string{
-	"gmail.com",
-	"yahoo.com",
-	"hotmail.com",
-	"outlook.com",
-	"aol.com",
-	"mail.com",
-	"yandex.ru",
-	"icloud.com",
-	"protonmail.com",
-	"zoho.com",
-	"gmx.com",
-	"hubspot.com",
-	"fastmail.com",
-	"hushmail.com",
-	"tutanota.com",
-	"hey.com",
-	"mail.ru",
-	"qq.com",
-	"163.com",
-	"126.com",
-	"sina.com",
-	"sohu.com",
-	"139.com",
-	"yeah.net",
-	"21cn.com",
-	"tom.com",
-	"sogou.com",
-	"live.com",
-	"msn.com",
-	"189.cn",
-	"263.net",
-	"aliyun.com",
-	"foxmail.com",
-	"vip.qq.com",
-	"vip.163.com",
-	"vip.126.com",
-	"vip.sina.com",
-	"vip.sohu.com",
-	"vip.139.com",
-	"vip.yeah.net",
-	"vip.21cn.com",
-	"vip.tom.com",
-	"vip.sogou.com",
-	"vip.foxmail.com",
-	"me.com",
-	"mac.com",
-	"yandex.com",
-	"rambler.ru",
-	"ya.ru",
-	"list.ru",
-	"inbox.ru",
-	"bk.ru",
-	"yandex.by",
-	"yandex.kz",
-	"yandex.ua",
-	"yandex.com.tr",
-	"yandex.com.am",
-	"yandex.az",
-	"yandex.co.il",
-	"yandex.com.ge",
-	"yandex.ee",
-	"yandex.fr",
-	"yandex.kg",
-	"yandex.lv",
-	"yandex.lt",
-	"yandex.md",
-	"yandex.tj",
-	"yandex.tm",
-	"yandex.uz",
-	"outlook.de",
-	"outlook.es",
-	"outlook.fr",
-	"outlook.it",
-	"outlook.jp",
-	"outlook.com.br",
-	"outlook.com.ar",
-	"outlook.com.mx",
-	"outlook.com.tr",
-	"outlook.com.au",
-	"outlook.com.sg",
-	"outlook.co.nz",
-	"outlook.co.id",
-	"outlook.co.th",
-	"outlook.ph",
-	"outlook.sa",
-	"outlook.ae",
-	"outlook.my",
-	"outlook.vn",
-	"outlook.cl",
-	"outlook.co.il",
-	"outlook.com.eg",
-	"outlook.com.pe",
-	"outlook.com.co",
-	"outlook.co.za",
-	"outlook.com.ng",
-	"outlook.com.pk",
-	"outlook.com.bd",
-	"outlook.com.hk",
-	"outlook.com.tw",
-	"outlook.com.cy",
-	"gmx.us",
-	"gmx.ca",
-	"gmx.fr",
-	"gmx.es",
-	"gmx.co.uk",
-	"gmx.ie",
-	"gmx.com.au",
-	"gmx.at",
-	"gmx.ch",
-	"gmx.net",
-	"gmx.it",
-	"gmx.nl",
-	"gmx.be",
-	"gmx.pt",
-	"gmx.se",
-	"gmx.fi",
-	"gmx.dk",
-	"gmx.no",
-	"gmx.co.nz",
-	"mail.com",
-	"email.com",
-	"usa.com",
-	"myself.com",
-	"consultant.com",
-	"post.com",
-	"europe.com",
-	"asia.com",
-	"iname.com",
-	"writeme.com",
-	"dr.com",
-	"engineer.com",
-	"cheerful.com",
-	"accountant.com",
-	"techie.com",
-	"linuxmail.org",
-	"uymail.com",
-	"contractor.net",
-	"activist.com",
-	"adexec.com",
-	"allergist.com",
-	"alumni.com",
-	"alumnidirector.com",
-	"angelic.com",
-	"appraiser.net",
-	"archaeologist.com",
-	"arcticmail.com",
-	"artlover.com",
-	"auctioneer.net",
-	"bartender.net",
-	"bikerider.com",
-	"birdlover.com",
-	"brew-master.com",
-	"cash4u.com",
-	"chef.net",
-	"chemist.com",
-	"clerk.com",
-	"clubmember.org",
-	"collector.org",
-	"columnist.com",
-	"comic.com",
-	"computer4u.com",
-	"consultant.com",
-	"contractor.com",
-	"coolsite.net",
-	"counsellor.com",
-	"cyberservices.com",
-	"deliveryman.com",
-	"diplomats.com",
-	"disposable.com",
-	"doctor.com",
-	"execs.com",
-	"fastservice.com",
-	"financier.com",
-	"fireman.net",
-	"gardener.com",
-	"geologist.com",
-	"graduate.org",
-	"graphic-designer.com",
-	"hairdresser.net",
-	"homemail.com",
-	"hot-shot.com",
-	"instruction.com",
-	"instructor.net",
-	"insurer.com",
-	"job4u.com",
-	"journalist.com",
-	"legislator.com",
-	"lobbyist.com",
-	"minister.com",
-	"musician.org",
-	"net-shopping.com",
-	"optician.com",
-	"orthodontist.net",
-	"pediatrician.com",
-	"photographer.net",
-	"physicist.net",
-	"planetmail.com",
-	"planetmail.net",
-	"politician.com",
-	"presidency.com",
-	"imap.gmail.com",
-	"outlook.office365.com",
-	"imap.mail.yahoo.com",
-	"imap.aol.com",
-	"imap.mail.me.com",
-	"imap.mail.com",
-	"imap.zoho.com",
-	"imap.yandex.com",
-	"imap.gmx.com",
-	"imaps.mail.ru",
-	"imap.mail.yandex.ru",
-	"imap.fastmail.com",
-	"imap.rambler.ru",
-	"imap.protonmail.com",
-	"imap.ukr.net",
-	"imap-mail.outlook.com",
-	"mail.gandi.net",
-	"imap.mail.uk",
-	"imap.rackspace.com",
-	"imap.hushmail.com",
-	"imap.earthlink.net",
-	"imap.cox.net",
-	"imap.mailbox.org",
-	"imap.namecheap.com",
-	"imap.pobox.com",
-	"mail.messagingengine.com",
-	"imap.suddenlink.net",
-	"imap.att.yahoo.com",
-	"imap.mail.com.tr",
-	"imap.mimecast.com",
-	"imap.emailsrvr.com",
-	"mail.zimbra.com",
-	"imap.tutanota.com",
-	"imap.hosted-exchange.com",
-	"imap.comcast.net",
-	"imap.posteo.de",
-	"mobile.charter.net",
-	"mail.hover.com",
-	"imap.mail.umn.edu",
-	"imap.windowslive.com",
-	"imap.xs4all.nl",
-	"imap.mindspring.com",
-	"imap.roadrunner.com",
-	"imap.softbank.jp",
-	"imap.xtra.co.nz",
-	"imap.web.de",
-	"imap.mail.lycos.com",
-	"imap.mail1.umn.edu",
-	"imap.lavabit.com",
-	"imap.optonline.net",
-	"imap.verizon.net",
-	"imap.mail.missouri.edu",
-	"outlook.office.com",
-	"imap.juno.com",
-	"imap.one.com",
-	"imap.rogers.com",
-	"imap.mail.naver.com",
-	"imap.netzero.net",
-	"imap.mail.com.cn",
-	"imap.mxhichina.com",
-	"imap.vodafone.it",
-	"imap.ziggo.nl",
-	"imap2.mail.ru",
-	"imap-mail.outlook.de",
-	"imap.otenet.gr",
-	"imap.ntlworld.com",
-	"imap.sbcglobal.net",
-	"imap.tiscali.it",
-	"mail.aa.net.uk",
-	"secureimap.t-online.de",
-	"imap.uol.com.br",
-	"imap.seznam.cz",
-	"imap.msn.com",
-	"imap.telenet.be",
-	"imap.netcologne.de",
-	"imap.iinet.net.au",
-	"imap.zonnet.nl",
-	"mail.binghamton.edu",
-	"imap.bellsouth.net",
-	"mail.gradcenter.cuny.edu",
-	"imap.swisscom.ch",
-	"imap.unitybox.de",
-	"imap.bluehost.com",
-	"imap.cfl.rr.com",
-	"imap.rcn.com",
-	"mail.pair.com",
-	"imap.optusnet.com.au",
-	"mail.westhost.com",
-	"imap.riseup.net",
-	"imap.shaw.ca",
-	"imap.eircom.net",
-	"imap.sina.com",
-	"imap.strato.de",
-	"imap.email.arizona.edu",
-	"imap.umich.edu",
-	"imap.uwm.edu",
-	"secureimap.ih.ku.dk",
-	"imap.vanderbilt.edu",
-	"imap.earthlink.com",
-	"secureimap.freenet.de",
-	"imap.jcom.home.ne.jp",
-	"imap.mweb.co.za",
-	"imap.tpg.com.au",
-	"imap.mail.unm.edu",
-	"secureimap.surf.net",
-	"imap.telfort.nl",
-	"imap.orange.fr",
-	"mail.web.cox.net",
-	"imap.swbell.net",
-	"imap.windstream.net",
-	"imap.centurylink.net",
-	"imap.carolina.rr.com",
-	"imap.frontier.com",
-	"imap.austin.rr.com",
-	"mail.tampabay.rr.com",
-	"imap.u.washington.edu",
-	"mail.rochester.edu",
-	"imap.hawaii.rr.com",
-	"imap.pacbell.net",
-	"imap.mail.usf.edu",
-	"imap.uci.edu",
-	"imap.sympatico.ca",
-	"mail.mayfirst.org",
-	"mail.buffalo.edu",
-	"imap.virginmedia.com",
-	"imap.umn.edu",
-	"imap.twc.com",
-	"imap.maine.rr.com",
-	"imap.sas.upenn.edu",
-	"imap.netvigator.com",
-	"imap.optimum.net",
-	"imap.blueyonder.co.uk",
-	"imap.xs4all.com",
-	"mail.nau.edu",
-	"imap.videotron.ca",
-	"imap.gmx.net",
-	"imap.bredband.net",
-	"imap.easynet.co.uk",
-	"secureimap.arcor.de",
-	"imap.telia.com",
-	"mail.monash.edu.au",
-	"imap.ufl.edu",
-	"imap.hanmail.net",
-	"mail.utexas.edu",
-	"imap.snu.ac.kr",
-	"imap.fsu.edu",
-	"imap.nyu.edu",
-	"mailhost.unc.edu",
-	"imap.wowway.com",
-	"imap.ptd.net",
-	"imap.ncsu.edu",
-	"imap.asu.edu",
-	"secureimap.1blu.de",
-	"imap.bigpond.net.au",
-	"imap.gatech.edu",
-	"imap.nau.edu",
-	"mail.radford.edu",
-	"imap.ou.edu",
-	"imap.hinet.net",
-	"webmail.utoronto.ca",
-	"imap.umn.edu",
-	"imap.cinci.rr.com",
-	"mail.ualberta.ca",
-	"imap.aliceposta.it",
-	"imap.iisc.ac.in",
-	"imap.ntlworld.ie",
-	"imap.louisville.edu",
-	"imap.unimelb.edu.au",
-	"imap.unsw.edu.au",
-	"imap.byu.edu",
-	"imap.syr.edu",
-	"imap.btopenworld.com",
-	"imap.drexel.edu",
-	"imap.vt.edu",
-	"imap.libero.it",
-	"imap.skynet.be",
-	"imap.asu.edu",
-	"imap.lsu.edu",
-	"imap.anu.edu.au",
-	"mail.iup.edu",
-	"imap.rug.nl",
-	"imap.du.edu",
-	"imap.ucsb.edu",
-	"imap.uark.edu",
-	"mail.unc.edu",
-	"imap.usc.edu",
-	"imap.metu.edu.tr",
-	"imap.psu.edu",
-	"imap.tcd.ie",
-	"imap.nd.edu",
-	"imap.ucdavis.edu",
-	"imap.indiana.edu",
-	"imap.ku.edu",
-	"imap.uoregon.edu",
-	"imap.yorku.ca",
-	"imap.ntu.edu.tw",
-	"imap.wisc.edu",
-	"imap.scarletmail.rutgers.edu",
-	"imap.uwaterloo.ca",
-	"imap.buffalo.edu",
-	"gmail.com",
-	"outlook.com",
-	"yahoo.com",
-	"mail.ru",
-	"qq.com",
-	"163.com",
-	"126.com",
-	"yeah.net",
-	"hotmail.com",
-	"icloud.com",
-	"naver.com",
-	"zoho.com",
-	"gmx.com",
-	"protonmail.com",
-	"yandex.ru",
-	"rambler.ru",
-	"daum.net",
-	"aol.com",
-	"fastmail.com",
-	"tutanota.com",
-	"gmx.de",
-	"web.de",
-	"runbox.com",
-	"mail.com",
-	"t-online.de",
-	"posteo.de",
-	"email.com",
-	"hushmail.com",
-	"msn.com",
-	"seznam.cz",
-	"ukr.net",
-	"inbox.lv",
-	"mail.bg",
-	"inbox.ru",
-	"mynet.com",
-	"rediffmail.com",
-	"wp.pl",
-	"o2.pl",
-	"interia.pl",
-	"tlen.pl",
-	"onet.pl",
-	"bluewin.ch",
-	"alice.it",
-	"tim.it",
-	"aruba.it",
-	"virgilio.it",
-	"poste.it",
-	"hispeed.ch",
-	"gmx.ch",
-	"freemail.hu",
-	"citromail.hu",
-	"mail.com",
-	"post.cz",
-	"volny.cz",
-	"freemail.cz",
-	"atlas.cz",
-	"centrum.cz",
-	"zoznam.sk",
-	"azet.sk",
-	"post.sk",
-	"abv.bg",
-	"dir.bg",
-	"mail.bg",
-	"i.ua",
-	"bigmir.net",
-	"bigpond.com",
-	"optusnet.com.au",
-	"live.com.au",
-	"gmail.com.au",
-	"telstra.com",
-	"iprimus.com.au",
-	"ozemail.com.au",
-	"webmail.co.za",
-	"mweb.co.za",
-	"telkomsa.net",
-	"afrihost.co.za",
-	"axxess.co.za",
-	"iburst.co.za",
-	"vodamail.co.za",
-	"bellnet.ca",
-	"shaw.ca",
-	"telus.net",
-	"videotron.ca",
-	"hotmail.ca",
-	"sympatico.ca",
-	"rogers.com",
-	"live.ca",
-	"cogeco.ca",
-	"sasktel.net",
-	"eastlink.ca",
-	"primus.ca",
-	"me.com",
-	"gmx.co.uk",
-	"talktalk.co.uk",
-	"verizon.net",
-	"comcast.net",
-	"charter.net",
-	"cox.net",
-	"sbcglobal.net",
-	"bellsouth.net",
-	"earthlink.net",
-	"hughes.net",
-	"ntlworld.com",
-	"o2.co.uk",
-	"orange.net",
-	"wanadoo.fr",
-	"free.fr",
-	"sfr.fr",
-	"neuf.fr",
-	"laposte.net",
-	"yahoo.fr",
-	"gmail.fr",
-	"outlook.fr",
-	"orange.fr",
-	"gmx.fr",
-	"caramail.fr",
-	"noos.fr",
-	"club-internet.fr",
-	"aliceadsl.fr",
-	"live.fr",
-	"hotmail.fr",
-	"aol.fr",
-	"t-online.hu",
-	"invitel.hu",
-	"vipmail.hu",
-	"upcmail.hu",
-	"digi.hu",
-	"indamail.hu",
-	"home.nl",
-	"ziggo.nl",
-	"hetnet.nl",
-	"planet.nl",
-	"casema.nl",
-	"gmail.com.tr",
-	"hotmail.com.tr",
-	"outlook.com.tr",
-	"yahoo.com.tr",
-	"mynet.com.tr",
-	"superonline.com",
-	"ttmail.com",
-	"turk.net",
-	"turkcell.com.tr",
-	"turktelekom.com.tr",
-	"d2.net.tr",
-	"ttnet.net.tr",
-	"kktcell.com",
-	"telenet.be",
-	"skynet.be",
-	"proximus.be",
-	"scarlet.be",
-	"voo.be",
-	"mail.be",
-	"pandora.be",
-	"gmail.com.br",
-	"hotmail.com.br",
-	"outlook.com.br",
-	"yahoo.com.br",
-	"uol.com.br",
-	"bol.com.br",
-	"ig.com.br",
-	"terra.com.br",
-	"itelefonica.com.br",
-	"r7.com",
-	"zipmail.com.br",
-	"globo.com",
-	"globomail.com",
-	"oi.com.br",
-	"ibest.com.br",
-	"vivo.com.br",
-	"live.com.ar",
-	"hotmail.com.ar",
-	"outlook.com.ar",
-	"yahoo.com.ar",
-	"gmail.com.ar",
-	"fibertel.com.ar",
-	"speedy.com.ar",
-	"arnet.com.ar",
-	"hotmail.com.mx",
-	"prodigy.net.mx",
-	"yahoo.com.mx",
-	"gmail.com.mx",
-	"live.com.mx",
-	"outlook.com.mx",
-	"att.net.mx",
-	"icloud.com.mx",
-	"telmex.net",
-	"prodigy.net.mx",
-	"infinitummail.com",
-	"mail.ru.nl",
-	"yandex.kz",
-	"mail.kz",
-	"inbox.kz",
-	"list.ru",
-	"bk.ru",
-	"xaker.ru",
-	"hotbox.ru",
-	"zmail.ru",
-	"pochta.ru",
-	"land.ru",
-	"newmail.ru",
-	"smtp.gmail.com",
-	"smtp.mail.yahoo.com",
-	"smtp.live.com",
-	"smtp.mail.me.com",
-	"smtp.mail.att.net",
-	"smtp.comcast.net",
-	"smtp.verizon.net",
-	"smtp.bizmail.yahoo.com",
-	"smtp.mail.aol.com",
-	"smtp.mail.com",
-	"smtp.zoho.com",
-	"smtp.fastmail.com",
-	"smtp.sendgrid.net",
-	"smtp.mailgun.org",
-	"smtp.postmarkapp.com",
-	"smtp.mandrillapp.com",
-	"smtp.sparkpostmail.com",
-	"smtp.mailjet.com",
-	"smtp.sendinblue.com",
-	"smtp.elasticemail.com",
-	"smtp.tipimail.com",
-	"smtp.pepipost.com",
-	"smtp.smtp2go.com",
-	"smtp.serversmtp.com",
-	"smtp.socketlabs.com",
-	"smtp.mailchannels.net",
-	"smtp.dreamhost.com",
-	"smtp.1and1.com",
-	"smtp.office365.com",
-	"smtp.mail.outlook.com",
-	"smtp.secureserver.net",
-	"smtp.earthlink.net",
-	"smtp.cox.net",
-	"smtp.earthlink.net",
-	"smtp.charter.net",
-	"smtp.mail.asu.edu",
-	"smtp.gmx.com",
-	"smtp.mail.ucla.edu",
-	"smtp.northwestern.edu",
-	"smtp.thesmtp.com",
-	"smtp.fatcow.com",
-	"smtp.ipage.com",
-	"smtp.hostgator.com",
-	"smtp.infomaniak.com",
-	"smtp.strato.de",
-	"smtp.rediffmail.com",
-	"smtp.hushmail.com",
-	"smtp.poczta.onet.pl",
-	"smtp.centrum.cz",
-	"smtp.rambler.ru",
-	"smtp.nic.ru",
-	"smtp.timeweb.ru",
-	"smtp.yandex.ru",
-	"smtp.mail.ru",
-	"smtp.mail.com.tr",
-	"smtp.aruba.it",
-	"smtp.aliyun.com",
-	"smtp.qq.com",
-	"smtp.21cn.com",
-	"smtp.126.com",
-	"smtp.yeah.net",
-	"smtp.tom.com",
-	"smtp.sina.com",
-	"smtp.sohu.com",
-	"smtp.139.com",
-	"smtp.189.cn",
-	"smtp.daum.net",
-	"smtp.hanmail.net",
-	"smtp.nate.com",
-	"smtp.naver.com",
-	"smtp.ziggo.nl",
-	"smtp.kpnmail.nl",
-	"smtp.xs4all.nl",
-	"smtp.upcmail.nl",
-	"smtp.telenet.be",
-	"smtp.skynet.be",
-	"smtp.proximus.be",
-	"smtp.orange.fr",
-	"smtp.sfr.fr",
-	"smtp.free.fr",
-	"smtp.laposte.net",
-	"smtp.bouygtel.fr",
-	"smtp.numericable.fr",
-	"smtp.wanadoo.fr",
-	"smtp.gmx.de",
-	"smtp.web.de",
-	"smtp.online.de",
-	"smtp.t-online.de",
-	"smtp.freenet.de",
-	"smtp.arcor.de",
-	"smtp.1und1.de",
-	"smtp.ziggo.nl",
-	"smtp.freenet.de",
-	"smtp.versatel.de",
-	"smtp.xs4all.nl",
-	"smtp.gmx.at",
-	"smtp.chello.at",
-	"smtp.inode.at",
-	"smtp.a1.net",
-	"smtp.bluewin.ch",
-	"smtp.gmx.ch",
-	"smtp.swissonline.ch",
-	"smtp.cablecom.ch",
-	"smtp.sunrise.ch",
-	"smtp.hispeed.ch",
-	"smtp.freesurf.ch",
-	"smtp.bigpond.com",
-	"smtp.Adam.com.au",
-	"smtp.iinet.net.au",
-	"smtp.optusnet.com.au",
-	"smtp.netspace.net.au",
-	"smtp.ozemail.com.au",
-	"smtp.exemail.com.au",
-	"smtp.iprimus.com.au",
-	"smtp.tpg.com.au",
-	"smtp.westnet.com.au",
-	"smtp.aapt.net.au",
-	"smtp.internode.on.net",
-	"smtp.smartchat.net.au",
-	"smtp.onthenet.com.au",
-	"smtp.mail2web.com",
-	"smtp.runbox.com",
-	"smtp.emailsrvr.com",
-	"smtp.megaline.com",
-	"smtp.xtra.co.nz",
-	"smtp.paradise.net.nz",
-	"smtp.clear.net.nz",
-	"smtp.vodafone.net.nz",
-	"smtp.orcon.net.nz",
-	"smtp.telstraclear.co.nz",
-	"smtp.slingshot.co.nz",
-	"smtp.ihug.co.nz",
-	"smtp.xnet.co.nz",
-	"smtp.actrix.co.nz",
-	"smtp.callplus.net.nz",
-	"smtp.eircom.net",
-	"smtp.indigo.ie",
-	"smtp.magnet.ie",
-	"smtp.utvinternet.com",
-	"smtp.ukonline.co.uk",
-	"smtp.virgin.net",
-	"smtp.ntlworld.com",
-	"smtp.blueyonder.co.uk",
-	"smtp.orangehome.co.uk",
-	"smtp.wanadoo.co.uk",
-	"smtp.freeserve.co.uk",
-	"smtp.fsmail.net",
-	"smtp.lineone.net",
-	"smtp.tiscali.co.uk",
-	"smtp.talktalk.net",
-	"smtp.rackspace.com",
-	"smtp.pair.com",
-	"smtp.name.com",
-	"smtp.zrinity.com",
-	"smtp.sherweb.com",
-	"smtp.netfirms.com",
-	"smtp.1and1.co.uk",
-	"smtp.one.com",
-	"smtp.gandi.net",
-	"smtp.ovh.net",
-	"smtp.online.net",
-	"smtp.yourisp.com",
-	"smtp.internetx.de",
-	"smtp.servage.net",
-	"smtp.unitedinternet.de",
-	"smtp.domainfactory.com",
-	"smtp.all-inkl.com",
-	"smtp.hetzner.de",
-	"smtp.df.eu",
-	"smtp.netcup.de",
-	"smtp.livemail.co.uk",
-	"smtp.123-reg.co.uk",
-	"smtp.fasthosts.co.uk",
-	"smtp.names.co.uk",
-	"smtp.easily.co.uk",
-	"smtp.namesco.net",
-	"smtp.mailtrap.io",
-	"smtp.email-mx.net",
-	"smtp.mailspons.com",
-	"smtp.socketlabs.com",
-	"smtp.siteground.eu",
-	"smtp.siteground.asia",
-	"smtp.flockmail.com",
-	"smtp.acelle.com",
-	"smtp.postal.io",
-	"smtp.turbo-smtp.com",
-	"smtp.inumbo.com",
-	"smtp.reacher.email",
-	"smtp.mailpace.com",
-	"smtp.interserver.net",
-	"smtp.tigertech.net",
-	"smtp.he.net",
-	"smtp.intermedia.net",
-	"smtp.liquidweb.com",
-	"smtp.greengeeks.com",
-	"smtp.a2hosting.com",
-	"smtp.arvixe.com",
-	"smtp.bluehost.com",
-	"smtp.site5.com",
-	"smtp.inmotionhosting.com",
-}
+// FIXME load from bit masked sni-mail.gz. We don't want any url fragments to be in the binary
+var mailHosts = []string{}
